crs {radiant.model}	R Documentation

Collaborative Filtering

Description

Collaborative Filtering

Usage

crs(dataset, id, prod, pred, rate, data_filter = "")

Arguments

dataset
Dataset

id
String with name of the variable containing user ids

prod
String with name of the variable with product ids

pred
Products to predict for

rate
String with name of the variable with product ratings

data_filter
Expression entered in, e.g., Data > View to filter the dataset in Radiant. The expression should be a string (e.g., "training == 1")

Details

See https://radiant-rstats.github.io/docs/model/crs.html for an example in Radiant

Value

A data.frame with the original data and a new column with predicted ratings

See Also

summary.crs to summarize results

plot.crs to plot results if the actual ratings are available

Examples

crs(ratings, id = "Users", prod = "Movies", pred = c("M6", "M7", "M8", "M9", "M10"),
    rate = "Ratings", data_filter = "training == 1") %>% str()

[Package radiant.model version 1.0.0 Index]