dtree {radiant.model}	R Documentation

Create a decision tree

Description

Create a decision tree

Usage

dtree(yl, opt = "max", base = character(0))

Arguments

yl
A yaml string or a list (e.g., from yaml::yaml.load_file())

opt
Find the maximum ("max") or minimum ("min") value for each decision node

base
List of variable definitions from a base tree used when calling a sub-tree

Details

See https://radiant-rstats.github.io/docs/model/dtree.html for an example in Radiant

Value

A list with the initial tree and the calculated tree

See Also

summary.dtree to summarize results

plot.dtree to plot results

sensitivity.dtree to plot results

Examples

yaml::as.yaml(movie_contract) %>% cat()
dtree(movie_contract, opt = "max") %>% summary(output = TRUE)

[Package radiant.model version 1.0.0 Index]