find_min {radiant.model}	R Documentation

Find minimum value of a vector

Description

Find minimum value of a vector

Usage

find_min(x, y)

Arguments

x
Variable to find the minimum for

y
Variable to find the value for at the maximum of var

Details

Find the value of y at the minimum value of x

Value

Value of val at the minimum of var

Examples

find_min(1:10, 21:30)

[Package radiant.model version 1.0.0 Index]