Gradient Boosted Trees using XGBoost

Gradient Boosted Trees using XGBoost

gbt(dataset, rvar, evar, type = "classification", lev = "",
  max_depth = 6, learning_rate = 0.3, min_split_loss = 0.01,
  min_child_weight = 1, subsample = 1, nrounds = 100,
  early_stopping_rounds = 3, wts = "None", seed = NA, check = "",
  data_filter = "", envir = parent.frame(), ...)

Arguments

dataset
Dataset

rvar
The response variable in the model

evar
Explanatory variables in the model

type
Model type (i.e., "classification" or "regression")

lev
Level to use as the first column in prediction output

max_depth
Maximum 'depth' of tree

learning_rate
Learning rate (eta)

min_split_loss
Minimal improvement (gamma)

min_child_weight
Minimum number of instances allowed in each node

subsample
Subsample ratio of the training instances (0-1)

nrounds
Number of trees to create

early_stopping_rounds
Early stopping rule

wts
Weights to use in estimation

seed
Random seed to use as the starting point

check
Optional estimation parameters

data_filter
Expression entered in, e.g., Data > View to filter the dataset in Radiant. The expression should be a string (e.g., "price > 10000")

envir
Environment to extract data from

...
Further arguments to pass to xgboost

Value

A list with all variables defined in gbt as an object of class gbt

Details

See https://radiant-rstats.github.io/docs/model/gbt.html for an example in Radiant

See also

Examples

gbt(titanic, "survived", c("pclass", "sex"), lev = "Yes") %>% summary()

#> Gradient Boosted Trees (XGBoost)
#> Type                 : Classification
#> Data                 : titanic
#> Response variable    : survived
#> Level                : Yes in survived
#> Explanatory variables: pclass, sex 
#> Max depth            : 6 
#> Learning rate (eta)  : 0.3 
#> Min split loss       : 0.01 
#> Min child weight     : 1 
#> Sub-sample           : 1 
#> Nr of rounds (trees) : 100 
#> Early stopping rounds: 3 
#> Nr obs               : 1 
#> 
#> Iteration history:
#> 
#> [1]	train-error:0.213806 
#> [2]	train-error:0.213806 
#> [3]	train-error:0.213806 
#> [4]	train-error:0.213806 
#> Stopping. Best iteration:
#> [1]	train-error:0.213806

gbt(titanic, "survived", c("pclass", "sex")) %>% str()

#> List of 23
#>  $ model                :List of 14
#>   ..$ handle         :Class 'xgb.Booster.handle' <externalptr> 
#>   ..$ raw            : raw [1:2556] 00 00 00 80 ...
#>   ..$ best_iteration : int 1
#>   ..$ best_ntreelimit: num 1
#>   ..$ best_score     : Named num 0.214
#>   .. ..- attr(*, "names")= chr "train-error"
#>   ..$ niter          : int 4
#>   ..$ evaluation_log :Classes ‘data.table’ and 'data.frame':	4 obs. of  2 variables:
#>   .. ..$ iter       : num [1:4] 1 2 3 4
#>   .. ..$ train_error: num [1:4] 0.214 0.214 0.214 0.214
#>   .. ..- attr(*, ".internal.selfref")=<externalptr> 
#>   ..$ call           : language xgb.train(params = params, data = dtrain, nrounds = nrounds, watchlist = watchlist,      verbose = verbose, print| __truncated__ ...
#>   ..$ params         :List of 7
#>   .. ..$ max_depth       : num 6
#>   .. ..$ learning_rate   : num 0.3
#>   .. ..$ min_split_loss  : num 0.01
#>   .. ..$ min_child_weight: num 1
#>   .. ..$ subsample       : num 1
#>   .. ..$ objective       : chr "binary:logistic"
#>   .. ..$ silent          : num 1
#>   ..$ callbacks      :List of 3
#>   .. ..$ cb.print.evaluation:function (env = parent.frame())  
#>   .. .. ..- attr(*, "call")= language cb.print.evaluation(period = print_every_n)
#>   .. .. ..- attr(*, "name")= chr "cb.print.evaluation"
#>   .. ..$ cb.evaluation.log  :function (env = parent.frame(), finalize = FALSE)  
#>   .. .. ..- attr(*, "call")= language cb.evaluation.log()
#>   .. .. ..- attr(*, "name")= chr "cb.evaluation.log"
#>   .. ..$ cb.early.stop      :function (env = parent.frame(), finalize = FALSE)  
#>   .. .. ..- attr(*, "call")= language cb.early.stop(stopping_rounds = early_stopping_rounds, maximize = maximize,      verbose = verbose)
#>   .. .. ..- attr(*, "name")= chr "cb.early.stop"
#>   ..$ feature_names  : chr [1:3] "pclass2nd" "pclass3rd" "sexmale"
#>   ..$ nfeatures      : int 3
#>   ..$ importance     :Classes ‘data.table’ and 'data.frame':	3 obs. of  4 variables:
#>   .. ..$ Feature  : chr [1:3] "sexmale" "pclass3rd" "pclass2nd"
#>   .. ..$ Gain     : num [1:3] 0.7464 0.219 0.0346
#>   .. ..$ Cover    : num [1:3] 0.418 0.418 0.165
#>   .. ..$ Frequency: num [1:3] 0.222 0.444 0.333
#>   .. ..- attr(*, ".internal.selfref")=<externalptr> 
#>   ..$ model          :Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	1043 obs. of  3 variables:
#>   .. ..$ survived: Factor w/ 2 levels "Yes","No": 1 1 2 2 2 1 1 2 1 2 ...
#>   .. ..$ pclass  : Factor w/ 3 levels "1st","2nd","3rd": 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..$ sex     : Factor w/ 2 levels "female","male": 1 2 1 2 1 2 1 2 1 2 ...
#>   .. ..- attr(*, "description")= chr "## Titanic\n\nThis dataset describes the survival status of individual passengers on the Titanic. The titanic d"| __truncated__
#>   ..- attr(*, "class")= chr "xgb.Booster"
#>  $ output               : chr [1:9] "[1]\ttrain-error:0.213806 " "Will train until train_error hasn't improved in 3 rounds." "" "[2]\ttrain-error:0.213806 " ...
#>  $ extra_args           : list()
#>  $ gbt_input            :List of 8
#>   ..$ max_depth            : num 6
#>   ..$ learning_rate        : num 0.3
#>   ..$ min_split_loss       : num 0.01
#>   ..$ nrounds              : num 100
#>   ..$ min_child_weight     : num 1
#>   ..$ subsample            : num 1
#>   ..$ early_stopping_rounds: num 3
#>   ..$ objective            : chr "binary:logistic"
#>  $ objective            : chr "binary:logistic"
#>  $ not_vary             : chr(0) 
#>  $ df_name              : chr "titanic"
#>  $ vars                 : chr [1:2] "pclass" "sex"
#>  $ rvar                 : chr "survived"
#>  $ evar                 : chr [1:2] "pclass" "sex"
#>  $ type                 : chr "classification"
#>  $ lev                  : chr "Yes"
#>  $ max_depth            : num 6
#>  $ learning_rate        : num 0.3
#>  $ min_split_loss       : num 0.01
#>  $ min_child_weight     : num 1
#>  $ subsample            : num 1
#>  $ nrounds              : num 100
#>  $ early_stopping_rounds: num 3
#>  $ wts                  : NULL
#>  $ seed                 : chr NA
#>  $ check                : chr ""
#>  $ data_filter          : chr ""
#>  - attr(*, "class")= chr [1:3] "gbt" "model" "list"

gbt(diamonds, "price", c("carat", "clarity"), type = "regression") %>% summary()

#> Gradient Boosted Trees (XGBoost)
#> Type                 : Regression
#> Data                 : diamonds
#> Response variable    : price
#> Explanatory variables: carat, clarity 
#> Max depth            : 6 
#> Learning rate (eta)  : 0.3 
#> Min split loss       : 0.01 
#> Min child weight     : 1 
#> Sub-sample           : 1 
#> Nr of rounds (trees) : 100 
#> Early stopping rounds: 3 
#> Nr obs               : 1 
#> 
#> Iteration history:
#> 
#> [1]	train-rmse:4004.123047 
#> [2]	train-rmse:2934.489990 
#> [3]	train-rmse:2209.100098 
#> [4]	train-rmse:1731.832153 
#> [5]	train-rmse:1429.542358 
#> [6]	train-rmse:1236.558594 
#> [7]	train-rmse:1112.986206 
#> [8]	train-rmse:1041.411255 
#> [9]	train-rmse:994.494995 
#> [10]	train-rmse:968.975891 
#> ...
#> [91]	train-rmse:747.641052 
#> [92]	train-rmse:747.167053 
#> [93]	train-rmse:745.880493 
#> [94]	train-rmse:744.933533 
#> [95]	train-rmse:744.060547 
#> [96]	train-rmse:743.978455 
#> [97]	train-rmse:743.872314 
#> [98]	train-rmse:743.621338 
#> [99]	train-rmse:743.458862 
#> [100]	train-rmse:743.281921