movie_contract {radiant.model}	R Documentation

Movie contract decision tree

Description

Movie contract decision tree

Usage

data(movie_contract)

Format

A nested list for decision and chance nodes, probabilities and payoffs

Details

Use decision analysis to create a decision tree for an actor facing a contract decision

[Package radiant.model version 1.0.0 Index]