plot.nb.predict {radiant.model}	R Documentation

Plot method for nb.predict function

Description

Plot method for nb.predict function

Usage

## S3 method for class 'nb.predict'
plot(x, xvar = "", facet_row = ".",
  facet_col = ".", color = ".class", ...)

Arguments

x
Return value from predict function predict.nb

xvar
Variable to display along the X-axis of the plot

facet_row
Create vertically arranged subplots for each level of the selected factor variable

facet_col
Create horizontally arranged subplots for each level of the selected factor variable

color
Adds color to a scatter plot to generate a heat map. For a line plot one line is created for each group and each is assigned a different color

...
further arguments passed to or from other methods

See Also

predict.nb to generate predictions

Examples

result <- nb(titanic, "survived", c("pclass", "sex", "age"))
pred <- predict(
  result,
  pred_cmd = c("pclass = levels(pclass)", "sex = levels(sex)", "age = seq(0, 100, 20)")
)
plot(pred, xvar = "age", facet_col = "sex", facet_row = "pclass")
pred <- predict(result, pred_data = titanic)
plot(pred, xvar = "age", facet_col = "sex")

[Package radiant.model version 1.0.0 Index]