plot.nn {radiant.model}	R Documentation

Plot method for the nn function

Description

Plot method for the nn function

Usage

## S3 method for class 'nn'
plot(x, plots = "garson", size = 12, nrobs = -1,
  shiny = FALSE, custom = FALSE, ...)

Arguments

x
Return value from nn

plots
Plots to produce for the specified Neural Network model. Use "" to avoid showing any plots (default). Options are "olden" or "garson" for importance plots, or "net" to depict the network structure

size
Font size used

nrobs
Number of data points to show in scatter plots (-1 for all)

shiny
Did the function call originate inside a shiny app

custom
Logical (TRUE, FALSE) to indicate if ggplot object (or list of ggplot objects) should be returned. This option can be used to customize plots (e.g., add a title, change x and y labels, etc.). See examples and http://docs.ggplot2.org for options.

...
further arguments passed to or from other methods

Details

See https://radiant-rstats.github.io/docs/model/nn.html for an example in Radiant

See Also

nn to generate results

summary.nn to summarize results

predict.nn for prediction

Examples

result <- nn(titanic, "survived", c("pclass", "sex"), lev = "Yes")
plot(result, plots = "net")
plot(result, plots = "olden")

[Package radiant.model version 1.0.0 Index]