plot.repeater {radiant.model}	R Documentation

Plot repeated simulation

Description

Plot repeated simulation

Usage

## S3 method for class 'repeater'
plot(x, bins = 20, shiny = FALSE, custom = FALSE,
  ...)

Arguments

x
Return value from repeater

bins
Number of bins used for histograms (1 - 50)

shiny
Did the function call originate inside a shiny app

custom
Logical (TRUE, FALSE) to indicate if ggplot object (or list of ggplot objects) should be returned. This option can be used to customize plots (e.g., add a title, change x and y labels, etc.). See examples and http://docs.ggplot2.org for options.

...
further arguments passed to or from other methods

See Also

repeater to run a repeated simulation

summary.repeater to summarize results from repeated simulation

[Package radiant.model version 1.0.0 Index]