predict.logistic {radiant.model}	R Documentation

Predict method for the logistic function

Description

Predict method for the logistic function

Usage

## S3 method for class 'logistic'
predict(object, pred_data = NULL, pred_cmd = "",
  conf_lev = 0.95, se = TRUE, interval = "confidence", dec = 3,
  ...)

Arguments

object
Return value from logistic

pred_data
Provide the dataframe to generate predictions (e.g., titanic). The dataset must contain all columns used in the estimation

pred_cmd
Generate predictions using a command. For example, ‘pclass = levels(pclass)' would produce predictions for the different levels of factor 'pclass'. To add another variable, create a vector of prediction strings, (e.g., c(’pclass = levels(pclass)', 'age = seq(0,100,20)')

conf_lev
Confidence level used to estimate confidence intervals (.95 is the default)

se
Logical that indicates if prediction standard errors should be calculated (default = FALSE)

interval
Type of interval calculation ("confidence" or "none"). Set to "none" if se is FALSE

dec
Number of decimals to show

...
further arguments passed to or from other methods

Details

See https://radiant-rstats.github.io/docs/model/logistic.html for an example in Radiant

See Also

logistic to generate the result

summary.logistic to summarize results

plot.logistic to plot results

plot.model.predict to plot prediction output

Examples

result <- logistic(titanic, "survived", c("pclass", "sex"), lev = "Yes")
  predict(result, pred_cmd = "pclass = levels(pclass)")
logistic(titanic, "survived", c("pclass", "sex"), lev = "Yes") %>%
  predict(pred_cmd = "sex = c('male','female')")
logistic(titanic, "survived", c("pclass", "sex"), lev = "Yes") %>%
  predict(pred_data = titanic)

[Package radiant.model version 1.0.0 Index]