Random Forest using Ranger

Random Forest using Ranger

rforest(dataset, rvar, evar, type = "classification", lev = "",
  mtry = NULL, num.trees = 100, min.node.size = 1,
  sample.fraction = 1, replace = NULL, wts = "None", seed = NA,
  check = "", data_filter = "", envir = parent.frame(), ...)

Arguments

dataset
Dataset

rvar
The response variable in the model

evar
Explanatory variables in the model

type
Model type (i.e., "classification" or "regression")

lev
Level to use as the first column in prediction output

mtry
Number of variables to possibly split at in each node. Default is the (rounded down) square root of the number variables

num.trees
Number of trees to create

min.node.size
Minimal node size

sample.fraction
Fraction of observations to sample. Default is 1 for sampling with replacement and 0.632 for sampling without replacement

replace
Sample with (TRUE) or without (FALSE) replacement. If replace is NULL it will be reset to TRUE if the sample.fraction is equal to 1 and will be set to FALSE otherwise

wts
Case weights to use in estimation

seed
Random seed to use as the starting point

check
Optional estimation parameters

data_filter
Expression entered in, e.g., Data > View to filter the dataset in Radiant. The expression should be a string (e.g., "price > 10000")

envir
Environment to extract data from

...
Further arguments to pass to ranger

Value

A list with all variables defined in rforest as an object of class rforest

Details

See https://radiant-rstats.github.io/docs/model/rforest.html for an example in Radiant

See also

Examples

rforest(titanic, "survived", c("pclass", "sex"), lev = "Yes") %>% summary()

#> Random Forest (Ranger)
#> Type                 : Classification
#> Data                 : titanic
#> Response variable    : survived
#> Level                : Yes in survived
#> Explanatory variables: pclass, sex 
#> Mtry                 : 
#> Number of trees      : 100 
#> Min node size        : 1 
#> Sample fraction      : 1 
#> Nr obs               : 1,043 
#> OOB prediction error : 0.154 

rforest(titanic, "survived", c("pclass", "sex")) %>% str()

#> List of 20
#>  $ model          :List of 15
#>   ..$ predictions              : num [1:1043, 1:2] 0.835 0.326 0.824 0.339 0.819 ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : chr [1:2] "Yes" "No"
#>   ..$ num.trees                : num 100
#>   ..$ num.independent.variables: num 2
#>   ..$ mtry                     : num 1
#>   ..$ min.node.size            : num 1
#>   ..$ variable.importance      : Named num [1:2] 0.0317 0.1204
#>   .. ..- attr(*, "names")= chr [1:2] "pclass" "sex"
#>   ..$ prediction.error         : num 0.155
#>   ..$ forest                   :List of 11
#>   .. ..$ dependent.varID           : num 0
#>   .. ..$ num.trees                 : num 100
#>   .. ..$ child.nodeIDs             :List of 100
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 0 3 0 5 0 0
#>   .. .. .. ..$ : num [1:7] 2 0 4 0 6 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:5] 1 0 3 0 0
#>   .. .. .. ..$ : num [1:5] 2 0 4 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:5] 1 0 3 0 0
#>   .. .. .. ..$ : num [1:5] 2 0 4 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:3] 1 0 0
#>   .. .. .. ..$ : num [1:3] 2 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:11] 1 3 5 0 0 7 9 0 0 0 ...
#>   .. .. .. ..$ : num [1:11] 2 4 6 0 0 8 10 0 0 0 ...
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:11] 1 3 5 7 0 0 9 0 0 0 ...
#>   .. .. .. ..$ : num [1:11] 2 4 6 8 0 0 10 0 0 0 ...
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:3] 1 0 0
#>   .. .. .. ..$ : num [1:3] 2 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 0 3 0 5 0 0
#>   .. .. .. ..$ : num [1:7] 2 0 4 0 6 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:9] 1 3 5 0 0 0 7 0 0
#>   .. .. .. ..$ : num [1:9] 2 4 6 0 0 0 8 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:5] 1 3 0 0 0
#>   .. .. .. ..$ : num [1:5] 2 4 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 0 3 0 5 0 0
#>   .. .. .. ..$ : num [1:7] 2 0 4 0 6 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:5] 1 0 3 0 0
#>   .. .. .. ..$ : num [1:5] 2 0 4 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 0 3 0 5 0 0
#>   .. .. .. ..$ : num [1:7] 2 0 4 0 6 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:3] 1 0 0
#>   .. .. .. ..$ : num [1:3] 2 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:11] 1 3 5 7 9 0 0 0 0 0 ...
#>   .. .. .. ..$ : num [1:11] 2 4 6 8 10 0 0 0 0 0 ...
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:3] 1 0 0
#>   .. .. .. ..$ : num [1:3] 2 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 3 5 0 0 0 0
#>   .. .. .. ..$ : num [1:7] 2 4 6 0 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:11] 1 3 5 7 9 0 0 0 0 0 ...
#>   .. .. .. ..$ : num [1:11] 2 4 6 8 10 0 0 0 0 0 ...
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:5] 1 0 3 0 0
#>   .. .. .. ..$ : num [1:5] 2 0 4 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:3] 1 0 0
#>   .. .. .. ..$ : num [1:3] 2 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:9] 1 3 5 0 7 0 0 0 0
#>   .. .. .. ..$ : num [1:9] 2 4 6 0 8 0 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:5] 1 3 0 0 0
#>   .. .. .. ..$ : num [1:5] 2 4 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 3 5 0 0 0 0
#>   .. .. .. ..$ : num [1:7] 2 4 6 0 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 3 0 5 0 0 0
#>   .. .. .. ..$ : num [1:7] 2 4 0 6 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 0 3 0 5 0 0
#>   .. .. .. ..$ : num [1:7] 2 0 4 0 6 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 0 3 0 5 0 0
#>   .. .. .. ..$ : num [1:7] 2 0 4 0 6 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 0 3 0 5 0 0
#>   .. .. .. ..$ : num [1:7] 2 0 4 0 6 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:9] 1 3 5 7 0 0 0 0 0
#>   .. .. .. ..$ : num [1:9] 2 4 6 8 0 0 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:9] 1 3 5 0 0 0 7 0 0
#>   .. .. .. ..$ : num [1:9] 2 4 6 0 0 0 8 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:5] 1 3 0 0 0
#>   .. .. .. ..$ : num [1:5] 2 4 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:9] 1 3 5 7 0 0 0 0 0
#>   .. .. .. ..$ : num [1:9] 2 4 6 8 0 0 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 0 3 0 5 0 0
#>   .. .. .. ..$ : num [1:7] 2 0 4 0 6 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:5] 1 3 0 0 0
#>   .. .. .. ..$ : num [1:5] 2 4 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:11] 1 3 5 7 0 0 9 0 0 0 ...
#>   .. .. .. ..$ : num [1:11] 2 4 6 8 0 0 10 0 0 0 ...
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:5] 1 3 0 0 0
#>   .. .. .. ..$ : num [1:5] 2 4 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:9] 1 3 5 7 0 0 0 0 0
#>   .. .. .. ..$ : num [1:9] 2 4 6 8 0 0 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:9] 1 0 3 5 7 0 0 0 0
#>   .. .. .. ..$ : num [1:9] 2 0 4 6 8 0 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 3 5 0 0 0 0
#>   .. .. .. ..$ : num [1:7] 2 4 6 0 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 3 0 5 0 0 0
#>   .. .. .. ..$ : num [1:7] 2 4 0 6 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:9] 1 3 5 7 0 0 0 0 0
#>   .. .. .. ..$ : num [1:9] 2 4 6 8 0 0 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 0 3 0 5 0 0
#>   .. .. .. ..$ : num [1:7] 2 0 4 0 6 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 0 3 5 0 0 0
#>   .. .. .. ..$ : num [1:7] 2 0 4 6 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 0 3 0 5 0 0
#>   .. .. .. ..$ : num [1:7] 2 0 4 0 6 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 0 3 0 5 0 0
#>   .. .. .. ..$ : num [1:7] 2 0 4 0 6 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:5] 1 3 0 0 0
#>   .. .. .. ..$ : num [1:5] 2 4 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:9] 1 3 5 7 0 0 0 0 0
#>   .. .. .. ..$ : num [1:9] 2 4 6 8 0 0 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 0 3 0 5 0 0
#>   .. .. .. ..$ : num [1:7] 2 0 4 0 6 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 3 0 0 5 0 0
#>   .. .. .. ..$ : num [1:7] 2 4 0 0 6 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 3 0 0 5 0 0
#>   .. .. .. ..$ : num [1:7] 2 4 0 0 6 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:11] 1 3 5 7 9 0 0 0 0 0 ...
#>   .. .. .. ..$ : num [1:11] 2 4 6 8 10 0 0 0 0 0 ...
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 0 3 0 5 0 0
#>   .. .. .. ..$ : num [1:7] 2 0 4 0 6 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:3] 1 0 0
#>   .. .. .. ..$ : num [1:3] 2 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:3] 1 0 0
#>   .. .. .. ..$ : num [1:3] 2 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 3 0 5 0 0 0
#>   .. .. .. ..$ : num [1:7] 2 4 0 6 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 3 0 5 0 0 0
#>   .. .. .. ..$ : num [1:7] 2 4 0 6 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:9] 1 3 5 0 0 0 7 0 0
#>   .. .. .. ..$ : num [1:9] 2 4 6 0 0 0 8 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 3 0 0 5 0 0
#>   .. .. .. ..$ : num [1:7] 2 4 0 0 6 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 3 5 0 0 0 0
#>   .. .. .. ..$ : num [1:7] 2 4 6 0 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 0 3 5 0 0 0
#>   .. .. .. ..$ : num [1:7] 2 0 4 6 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:11] 1 3 5 0 0 7 9 0 0 0 ...
#>   .. .. .. ..$ : num [1:11] 2 4 6 0 0 8 10 0 0 0 ...
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 3 5 0 0 0 0
#>   .. .. .. ..$ : num [1:7] 2 4 6 0 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 0 3 0 5 0 0
#>   .. .. .. ..$ : num [1:7] 2 0 4 0 6 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 3 5 0 0 0 0
#>   .. .. .. ..$ : num [1:7] 2 4 6 0 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:9] 1 3 5 0 0 0 7 0 0
#>   .. .. .. ..$ : num [1:9] 2 4 6 0 0 0 8 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 3 5 0 0 0 0
#>   .. .. .. ..$ : num [1:7] 2 4 6 0 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:5] 1 3 0 0 0
#>   .. .. .. ..$ : num [1:5] 2 4 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:5] 1 3 0 0 0
#>   .. .. .. ..$ : num [1:5] 2 4 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:9] 1 3 5 0 7 0 0 0 0
#>   .. .. .. ..$ : num [1:9] 2 4 6 0 8 0 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:3] 1 0 0
#>   .. .. .. ..$ : num [1:3] 2 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:3] 1 0 0
#>   .. .. .. ..$ : num [1:3] 2 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:3] 1 0 0
#>   .. .. .. ..$ : num [1:3] 2 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:9] 1 3 5 7 0 0 0 0 0
#>   .. .. .. ..$ : num [1:9] 2 4 6 8 0 0 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:9] 1 3 5 7 0 0 0 0 0
#>   .. .. .. ..$ : num [1:9] 2 4 6 8 0 0 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:3] 1 0 0
#>   .. .. .. ..$ : num [1:3] 2 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:5] 1 0 3 0 0
#>   .. .. .. ..$ : num [1:5] 2 0 4 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:3] 1 0 0
#>   .. .. .. ..$ : num [1:3] 2 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 3 0 5 0 0 0
#>   .. .. .. ..$ : num [1:7] 2 4 0 6 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 3 5 0 0 0 0
#>   .. .. .. ..$ : num [1:7] 2 4 6 0 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:11] 1 3 5 7 9 0 0 0 0 0 ...
#>   .. .. .. ..$ : num [1:11] 2 4 6 8 10 0 0 0 0 0 ...
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 3 5 0 0 0 0
#>   .. .. .. ..$ : num [1:7] 2 4 6 0 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:5] 1 0 3 0 0
#>   .. .. .. ..$ : num [1:5] 2 0 4 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:9] 1 3 0 5 7 0 0 0 0
#>   .. .. .. ..$ : num [1:9] 2 4 0 6 8 0 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:9] 1 3 5 7 0 0 0 0 0
#>   .. .. .. ..$ : num [1:9] 2 4 6 8 0 0 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:9] 1 3 5 0 0 0 7 0 0
#>   .. .. .. ..$ : num [1:9] 2 4 6 0 0 0 8 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 0 3 5 0 0 0
#>   .. .. .. ..$ : num [1:7] 2 0 4 6 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:5] 1 3 0 0 0
#>   .. .. .. ..$ : num [1:5] 2 4 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:3] 1 0 0
#>   .. .. .. ..$ : num [1:3] 2 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 3 0 5 0 0 0
#>   .. .. .. ..$ : num [1:7] 2 4 0 6 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:11] 1 3 5 7 0 0 9 0 0 0 ...
#>   .. .. .. ..$ : num [1:11] 2 4 6 8 0 0 10 0 0 0 ...
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 0 3 0 5 0 0
#>   .. .. .. ..$ : num [1:7] 2 0 4 0 6 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 3 0 5 0 0 0
#>   .. .. .. ..$ : num [1:7] 2 4 0 6 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:9] 1 0 3 5 7 0 0 0 0
#>   .. .. .. ..$ : num [1:9] 2 0 4 6 8 0 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:5] 1 3 0 0 0
#>   .. .. .. ..$ : num [1:5] 2 4 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 3 5 0 0 0 0
#>   .. .. .. ..$ : num [1:7] 2 4 6 0 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 0 3 0 5 0 0
#>   .. .. .. ..$ : num [1:7] 2 0 4 0 6 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 0 3 5 0 0 0
#>   .. .. .. ..$ : num [1:7] 2 0 4 6 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:7] 1 3 5 0 0 0 0
#>   .. .. .. ..$ : num [1:7] 2 4 6 0 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:9] 1 3 5 7 0 0 0 0 0
#>   .. .. .. ..$ : num [1:9] 2 4 6 8 0 0 0 0 0
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ : num [1:3] 1 0 0
#>   .. .. .. ..$ : num [1:3] 2 0 0
#>   .. .. .. [list output truncated]
#>   .. ..$ split.varIDs              :List of 100
#>   .. .. ..$ : num [1:7] 1 0 2 0 1 0 0
#>   .. .. ..$ : num [1:5] 2 0 1 0 0
#>   .. .. ..$ : num [1:5] 2 0 1 0 0
#>   .. .. ..$ : num [1:3] 2 0 0
#>   .. .. ..$ : num [1:11] 1 2 2 0 0 1 1 0 0 0 ...
#>   .. .. ..$ : num [1:11] 2 1 1 1 0 0 1 0 0 0 ...
#>   .. .. ..$ : num [1:3] 2 0 0
#>   .. .. ..$ : num [1:7] 1 0 1 0 2 0 0
#>   .. .. ..$ : num [1:9] 1 2 1 0 0 0 2 0 0
#>   .. .. ..$ : num [1:5] 2 1 0 0 0
#>   .. .. ..$ : num [1:7] 2 0 1 0 1 0 0
#>   .. .. ..$ : num [1:5] 2 0 1 0 0
#>   .. .. ..$ : num [1:7] 1 0 2 0 1 0 0
#>   .. .. ..$ : num [1:3] 2 0 0
#>   .. .. ..$ : num [1:11] 1 1 2 2 2 0 0 0 0 0 ...
#>   .. .. ..$ : num [1:3] 2 0 0
#>   .. .. ..$ : num [1:7] 1 2 1 0 0 0 0
#>   .. .. ..$ : num [1:11] 1 1 2 2 2 0 0 0 0 0 ...
#>   .. .. ..$ : num [1:5] 2 0 1 0 0
#>   .. .. ..$ : num [1:3] 2 0 0
#>   .. .. ..$ : num [1:9] 1 1 2 0 2 0 0 0 0
#>   .. .. ..$ : num [1:5] 2 1 0 0 0
#>   .. .. ..$ : num [1:7] 2 1 1 0 0 0 0
#>   .. .. ..$ : num [1:7] 2 1 0 1 0 0 0
#>   .. .. ..$ : num [1:7] 2 0 1 0 1 0 0
#>   .. .. ..$ : num [1:7] 2 0 1 0 1 0 0
#>   .. .. ..$ : num [1:7] 2 0 1 0 1 0 0
#>   .. .. ..$ : num [1:9] 1 1 2 2 0 0 0 0 0
#>   .. .. ..$ : num [1:9] 1 2 2 0 0 0 1 0 0
#>   .. .. ..$ : num [1:5] 2 1 0 0 0
#>   .. .. ..$ : num [1:9] 2 1 1 1 0 0 0 0 0
#>   .. .. ..$ : num [1:7] 2 0 1 0 1 0 0
#>   .. .. ..$ : num [1:5] 2 1 0 0 0
#>   .. .. ..$ : num [1:11] 2 1 1 1 0 0 1 0 0 0 ...
#>   .. .. ..$ : num [1:5] 2 1 0 0 0
#>   .. .. ..$ : num [1:9] 2 1 1 1 0 0 0 0 0
#>   .. .. ..$ : num [1:9] 1 0 1 2 2 0 0 0 0
#>   .. .. ..$ : num [1:7] 2 1 1 0 0 0 0
#>   .. .. ..$ : num [1:7] 1 2 0 1 0 0 0
#>   .. .. ..$ : num [1:9] 1 2 2 1 0 0 0 0 0
#>   .. .. ..$ : num [1:7] 2 0 1 0 1 0 0
#>   .. .. ..$ : num [1:7] 1 0 1 2 0 0 0
#>   .. .. ..$ : num [1:7] 2 0 1 0 1 0 0
#>   .. .. ..$ : num [1:7] 2 0 1 0 1 0 0
#>   .. .. ..$ : num [1:5] 2 1 0 0 0
#>   .. .. ..$ : num [1:9] 2 1 1 1 0 0 0 0 0
#>   .. .. ..$ : num [1:7] 1 0 2 0 1 0 0
#>   .. .. ..$ : num [1:7] 1 2 0 0 1 0 0
#>   .. .. ..$ : num [1:7] 1 1 0 0 2 0 0
#>   .. .. ..$ : num [1:11] 1 2 2 1 1 0 0 0 0 0 ...
#>   .. .. ..$ : num [1:7] 2 0 1 0 1 0 0
#>   .. .. ..$ : num [1:3] 2 0 0
#>   .. .. ..$ : num [1:3] 2 0 0
#>   .. .. ..$ : num [1:7] 2 1 0 1 0 0 0
#>   .. .. ..$ : num [1:7] 2 1 0 1 0 0 0
#>   .. .. ..$ : num [1:9] 2 1 1 0 0 0 1 0 0
#>   .. .. ..$ : num [1:7] 1 1 0 0 2 0 0
#>   .. .. ..$ : num [1:7] 1 2 2 0 0 0 0
#>   .. .. ..$ : num [1:7] 1 0 2 1 0 0 0
#>   .. .. ..$ : num [1:11] 1 2 2 0 0 1 1 0 0 0 ...
#>   .. .. ..$ : num [1:7] 1 2 2 0 0 0 0
#>   .. .. ..$ : num [1:7] 1 0 2 0 1 0 0
#>   .. .. ..$ : num [1:7] 1 2 1 0 0 0 0
#>   .. .. ..$ : num [1:9] 2 1 1 0 0 0 1 0 0
#>   .. .. ..$ : num [1:7] 1 1 2 0 0 0 0
#>   .. .. ..$ : num [1:5] 2 1 0 0 0
#>   .. .. ..$ : num [1:5] 2 1 0 0 0
#>   .. .. ..$ : num [1:9] 1 1 2 0 2 0 0 0 0
#>   .. .. ..$ : num [1:3] 2 0 0
#>   .. .. ..$ : num [1:3] 2 0 0
#>   .. .. ..$ : num [1:3] 2 0 0
#>   .. .. ..$ : num [1:9] 1 1 2 2 0 0 0 0 0
#>   .. .. ..$ : num [1:9] 2 1 1 1 0 0 0 0 0
#>   .. .. ..$ : num [1:3] 2 0 0
#>   .. .. ..$ : num [1:5] 1 0 2 0 0
#>   .. .. ..$ : num [1:3] 2 0 0
#>   .. .. ..$ : num [1:7] 2 1 0 1 0 0 0
#>   .. .. ..$ : num [1:7] 2 1 1 0 0 0 0
#>   .. .. ..$ : num [1:11] 1 2 2 1 1 0 0 0 0 0 ...
#>   .. .. ..$ : num [1:7] 1 2 2 0 0 0 0
#>   .. .. ..$ : num [1:5] 2 0 1 0 0
#>   .. .. ..$ : num [1:9] 1 2 0 1 1 0 0 0 0
#>   .. .. ..$ : num [1:9] 1 2 2 1 0 0 0 0 0
#>   .. .. ..$ : num [1:9] 2 1 1 0 0 0 1 0 0
#>   .. .. ..$ : num [1:7] 1 0 2 1 0 0 0
#>   .. .. ..$ : num [1:5] 2 1 0 0 0
#>   .. .. ..$ : num [1:3] 2 0 0
#>   .. .. ..$ : num [1:7] 2 1 0 1 0 0 0
#>   .. .. ..$ : num [1:11] 2 1 1 1 0 0 1 0 0 0 ...
#>   .. .. ..$ : num [1:7] 2 0 1 0 1 0 0
#>   .. .. ..$ : num [1:7] 2 1 0 1 0 0 0
#>   .. .. ..$ : num [1:9] 1 0 2 1 1 0 0 0 0
#>   .. .. ..$ : num [1:5] 2 1 0 0 0
#>   .. .. ..$ : num [1:7] 2 1 1 0 0 0 0
#>   .. .. ..$ : num [1:7] 2 0 1 0 1 0 0
#>   .. .. ..$ : num [1:7] 1 0 1 2 0 0 0
#>   .. .. ..$ : num [1:7] 1 1 2 0 0 0 0
#>   .. .. ..$ : num [1:9] 2 1 1 1 0 0 0 0 0
#>   .. .. ..$ : num [1:3] 2 0 0
#>   .. .. .. [list output truncated]
#>   .. ..$ split.values              :List of 100
#>   .. .. ..$ : num [1:7] 1.5 0 1.5 0 2.5 0 0
#>   .. .. ..$ : num [1:5] 1.5 0 1.5 0 0
#>   .. .. ..$ : num [1:5] 1.5 0 1.5 0 0
#>   .. .. ..$ : num [1:3] 1.5 0 0
#>   .. .. ..$ : num [1:11] 1.5 1.5 1.5 0 0 2.5 2.5 0 0 0 ...
#>   .. .. ..$ : num [1:11] 1.5 2.5 1.5 1.5 0 0 2.5 0 0 0 ...
#>   .. .. ..$ : num [1:3] 1.5 0 0
#>   .. .. ..$ : num [1:7] 1.5 0 2.5 0 1.5 0 0
#>   .. .. ..$ : num [1:9] 1.5 1.5 2.5 0 0 0 1.5 0 0
#>   .. .. ..$ : num [1:5] 1.5 2.5 0 0 0
#>   .. .. ..$ : num [1:7] 1.5 0 1.5 0 2.5 0 0
#>   .. .. ..$ : num [1:5] 1.5 0 1.5 0 0
#>   .. .. ..$ : num [1:7] 1.5 0 1.5 0 2.5 0 0
#>   .. .. ..$ : num [1:3] 1.5 0 0
#>   .. .. ..$ : num [1:11] 2.5 1.5 1.5 1.5 1.5 0 0 0 0 0 ...
#>   .. .. ..$ : num [1:3] 1.5 0 0
#>   .. .. ..$ : num [1:7] 1.5 1.5 2.5 0 0 0 0
#>   .. .. ..$ : num [1:11] 2.5 1.5 1.5 1.5 1.5 0 0 0 0 0 ...
#>   .. .. ..$ : num [1:5] 1.5 0 1.5 0 0
#>   .. .. ..$ : num [1:3] 1.5 0 0
#>   .. .. ..$ : num [1:9] 2.5 1.5 1.5 0 1.5 0 0 0 0
#>   .. .. ..$ : num [1:5] 1.5 2.5 0 0 0
#>   .. .. ..$ : num [1:7] 1.5 2.5 1.5 0 0 0 0
#>   .. .. ..$ : num [1:7] 1.5 2.5 0 1.5 0 0 0
#>   .. .. ..$ : num [1:7] 1.5 0 1.5 0 2.5 0 0
#>   .. .. ..$ : num [1:7] 1.5 0 1.5 0 2.5 0 0
#>   .. .. ..$ : num [1:7] 1.5 0 1.5 0 2.5 0 0
#>   .. .. ..$ : num [1:9] 2.5 1.5 1.5 1.5 0 0 0 0 0
#>   .. .. ..$ : num [1:9] 1.5 1.5 1.5 0 0 0 2.5 0 0
#>   .. .. ..$ : num [1:5] 1.5 2.5 0 0 0
#>   .. .. ..$ : num [1:9] 1.5 2.5 1.5 1.5 0 0 0 0 0
#>   .. .. ..$ : num [1:7] 1.5 0 1.5 0 2.5 0 0
#>   .. .. ..$ : num [1:5] 1.5 2.5 0 0 0
#>   .. .. ..$ : num [1:11] 1.5 2.5 1.5 1.5 0 0 2.5 0 0 0 ...
#>   .. .. ..$ : num [1:5] 1.5 2.5 0 0 0
#>   .. .. ..$ : num [1:9] 1.5 2.5 1.5 1.5 0 0 0 0 0
#>   .. .. ..$ : num [1:9] 1.5 0 2.5 1.5 1.5 0 0 0 0
#>   .. .. ..$ : num [1:7] 1.5 2.5 1.5 0 0 0 0
#>   .. .. ..$ : num [1:7] 2.5 1.5 0 1.5 0 0 0
#>   .. .. ..$ : num [1:9] 2.5 1.5 1.5 1.5 0 0 0 0 0
#>   .. .. ..$ : num [1:7] 1.5 0 1.5 0 2.5 0 0
#>   .. .. ..$ : num [1:7] 1.5 0 2.5 1.5 0 0 0
#>   .. .. ..$ : num [1:7] 1.5 0 1.5 0 2.5 0 0
#>   .. .. ..$ : num [1:7] 1.5 0 1.5 0 2.5 0 0
#>   .. .. ..$ : num [1:5] 1.5 2.5 0 0 0
#>   .. .. ..$ : num [1:9] 1.5 2.5 1.5 1.5 0 0 0 0 0
#>   .. .. ..$ : num [1:7] 1.5 0 1.5 0 2.5 0 0
#>   .. .. ..$ : num [1:7] 2.5 1.5 0 0 1.5 0 0
#>   .. .. ..$ : num [1:7] 2.5 1.5 0 0 1.5 0 0
#>   .. .. ..$ : num [1:11] 2.5 1.5 1.5 1.5 1.5 0 0 0 0 0 ...
#>   .. .. ..$ : num [1:7] 1.5 0 1.5 0 2.5 0 0
#>   .. .. ..$ : num [1:3] 1.5 0 0
#>   .. .. ..$ : num [1:3] 1.5 0 0
#>   .. .. ..$ : num [1:7] 1.5 2.5 0 1.5 0 0 0
#>   .. .. ..$ : num [1:7] 1.5 2.5 0 1.5 0 0 0
#>   .. .. ..$ : num [1:9] 1.5 2.5 1.5 0 0 0 2.5 0 0
#>   .. .. ..$ : num [1:7] 2.5 1.5 0 0 1.5 0 0
#>   .. .. ..$ : num [1:7] 2.5 1.5 1.5 0 0 0 0
#>   .. .. ..$ : num [1:7] 1.5 0 1.5 2.5 0 0 0
#>   .. .. ..$ : num [1:11] 1.5 1.5 1.5 0 0 2.5 2.5 0 0 0 ...
#>   .. .. ..$ : num [1:7] 2.5 1.5 1.5 0 0 0 0
#>   .. .. ..$ : num [1:7] 1.5 0 1.5 0 2.5 0 0
#>   .. .. ..$ : num [1:7] 1.5 1.5 2.5 0 0 0 0
#>   .. .. ..$ : num [1:9] 1.5 2.5 1.5 0 0 0 2.5 0 0
#>   .. .. ..$ : num [1:7] 2.5 1.5 1.5 0 0 0 0
#>   .. .. ..$ : num [1:5] 1.5 2.5 0 0 0
#>   .. .. ..$ : num [1:5] 1.5 2.5 0 0 0
#>   .. .. ..$ : num [1:9] 2.5 1.5 1.5 0 1.5 0 0 0 0
#>   .. .. ..$ : num [1:3] 1.5 0 0
#>   .. .. ..$ : num [1:3] 1.5 0 0
#>   .. .. ..$ : num [1:3] 1.5 0 0
#>   .. .. ..$ : num [1:9] 2.5 1.5 1.5 1.5 0 0 0 0 0
#>   .. .. ..$ : num [1:9] 1.5 2.5 1.5 1.5 0 0 0 0 0
#>   .. .. ..$ : num [1:3] 1.5 0 0
#>   .. .. ..$ : num [1:5] 1.5 0 1.5 0 0
#>   .. .. ..$ : num [1:3] 1.5 0 0
#>   .. .. ..$ : num [1:7] 1.5 2.5 0 1.5 0 0 0
#>   .. .. ..$ : num [1:7] 1.5 2.5 1.5 0 0 0 0
#>   .. .. ..$ : num [1:11] 2.5 1.5 1.5 1.5 1.5 0 0 0 0 0 ...
#>   .. .. ..$ : num [1:7] 2.5 1.5 1.5 0 0 0 0
#>   .. .. ..$ : num [1:5] 1.5 0 1.5 0 0
#>   .. .. ..$ : num [1:9] 2.5 1.5 0 1.5 1.5 0 0 0 0
#>   .. .. ..$ : num [1:9] 2.5 1.5 1.5 1.5 0 0 0 0 0
#>   .. .. ..$ : num [1:9] 1.5 2.5 1.5 0 0 0 2.5 0 0
#>   .. .. ..$ : num [1:7] 1.5 0 1.5 2.5 0 0 0
#>   .. .. ..$ : num [1:5] 1.5 2.5 0 0 0
#>   .. .. ..$ : num [1:3] 1.5 0 0
#>   .. .. ..$ : num [1:7] 1.5 2.5 0 1.5 0 0 0
#>   .. .. ..$ : num [1:11] 1.5 2.5 1.5 1.5 0 0 2.5 0 0 0 ...
#>   .. .. ..$ : num [1:7] 1.5 0 1.5 0 2.5 0 0
#>   .. .. ..$ : num [1:7] 1.5 2.5 0 1.5 0 0 0
#>   .. .. ..$ : num [1:9] 1.5 0 1.5 2.5 2.5 0 0 0 0
#>   .. .. ..$ : num [1:5] 1.5 2.5 0 0 0
#>   .. .. ..$ : num [1:7] 1.5 2.5 1.5 0 0 0 0
#>   .. .. ..$ : num [1:7] 1.5 0 1.5 0 2.5 0 0
#>   .. .. ..$ : num [1:7] 1.5 0 2.5 1.5 0 0 0
#>   .. .. ..$ : num [1:7] 2.5 1.5 1.5 0 0 0 0
#>   .. .. ..$ : num [1:9] 1.5 2.5 1.5 1.5 0 0 0 0 0
#>   .. .. ..$ : num [1:3] 1.5 0 0
#>   .. .. .. [list output truncated]
#>   .. ..$ is.ordered                : logi [1:3] TRUE TRUE TRUE
#>   .. ..$ class.values              : num [1:2] 1 2
#>   .. ..$ terminal.class.counts     :List of 100
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.634 0.366
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.592 0.408
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.155 0.845
#>   .. .. .. ..$ : num [1:2] 0.168 0.832
#>   .. .. ..$ :List of 5
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.771 0.229
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.337 0.663
#>   .. .. .. ..$ : num [1:2] 0.162 0.838
#>   .. .. ..$ :List of 5
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.775 0.225
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.384 0.616
#>   .. .. .. ..$ : num [1:2] 0.159 0.841
#>   .. .. ..$ :List of 3
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.748 0.252
#>   .. .. .. ..$ : num [1:2] 0.2 0.8
#>   .. .. ..$ :List of 11
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.963 0.037
#>   .. .. .. ..$ : num [1:2] 0.344 0.656
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.9053 0.0947
#>   .. .. .. ..$ : num [1:2] 0.429 0.571
#>   .. .. .. ..$ : num [1:2] 0.173 0.827
#>   .. .. .. ..$ : num [1:2] 0.19 0.81
#>   .. .. ..$ :List of 11
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.429 0.571
#>   .. .. .. ..$ : num [1:2] 0.453 0.547
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.944 0.056
#>   .. .. .. ..$ : num [1:2] 0.9062 0.0938
#>   .. .. .. ..$ : num [1:2] 0.146 0.854
#>   .. .. .. ..$ : num [1:2] 0.16 0.84
#>   .. .. ..$ :List of 3
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.774 0.226
#>   .. .. .. ..$ : num [1:2] 0.22 0.78
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.645 0.355
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.398 0.602
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.463 0.537
#>   .. .. .. ..$ : num [1:2] 0.189 0.811
#>   .. .. ..$ :List of 9
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.9722 0.0278
#>   .. .. .. ..$ : num [1:2] 0.326 0.674
#>   .. .. .. ..$ : num [1:2] 0.432 0.568
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.503 0.497
#>   .. .. .. ..$ : num [1:2] 0.167 0.833
#>   .. .. ..$ :List of 5
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.211 0.789
#>   .. .. .. ..$ : num [1:2] 0.9714 0.0286
#>   .. .. .. ..$ : num [1:2] 0.434 0.566
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.738 0.262
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.343 0.657
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.146 0.854
#>   .. .. .. ..$ : num [1:2] 0.132 0.868
#>   .. .. ..$ :List of 5
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.774 0.226
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.287 0.713
#>   .. .. .. ..$ : num [1:2] 0.175 0.825
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.637 0.363
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.63 0.37
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.12 0.88
#>   .. .. .. ..$ : num [1:2] 0.178 0.822
#>   .. .. ..$ :List of 3
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.73 0.27
#>   .. .. .. ..$ : num [1:2] 0.202 0.798
#>   .. .. ..$ :List of 11
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.429 0.571
#>   .. .. .. ..$ : num [1:2] 0.195 0.805
#>   .. .. .. ..$ : num [1:2] 0.9739 0.0261
#>   .. .. .. ..$ : num [1:2] 0.313 0.687
#>   .. .. .. ..$ : num [1:2] 0.817 0.183
#>   .. .. .. ..$ : num [1:2] 0.187 0.813
#>   .. .. ..$ :List of 3
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.741 0.259
#>   .. .. .. ..$ : num [1:2] 0.193 0.807
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.9706 0.0294
#>   .. .. .. ..$ : num [1:2] 0.371 0.629
#>   .. .. .. ..$ : num [1:2] 0.454 0.546
#>   .. .. .. ..$ : num [1:2] 0.29 0.71
#>   .. .. ..$ :List of 11
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.529 0.471
#>   .. .. .. ..$ : num [1:2] 0.2 0.8
#>   .. .. .. ..$ : num [1:2] 0.9658 0.0342
#>   .. .. .. ..$ : num [1:2] 0.329 0.671
#>   .. .. .. ..$ : num [1:2] 0.9495 0.0505
#>   .. .. .. ..$ : num [1:2] 0.144 0.856
#>   .. .. ..$ :List of 5
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.757 0.243
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.293 0.707
#>   .. .. .. ..$ : num [1:2] 0.194 0.806
#>   .. .. ..$ :List of 3
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.749 0.251
#>   .. .. .. ..$ : num [1:2] 0.191 0.809
#>   .. .. ..$ :List of 9
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.629 0.371
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.446 0.554
#>   .. .. .. ..$ : num [1:2] 0.185 0.815
#>   .. .. .. ..$ : num [1:2] 0.847 0.153
#>   .. .. .. ..$ : num [1:2] 0.184 0.816
#>   .. .. ..$ :List of 5
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.199 0.801
#>   .. .. .. ..$ : num [1:2] 0.9528 0.0472
#>   .. .. .. ..$ : num [1:2] 0.465 0.535
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.9598 0.0402
#>   .. .. .. ..$ : num [1:2] 0.436 0.564
#>   .. .. .. ..$ : num [1:2] 0.423 0.577
#>   .. .. .. ..$ : num [1:2] 0.17 0.83
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.225 0.775
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.469 0.531
#>   .. .. .. ..$ : num [1:2] 0.947 0.053
#>   .. .. .. ..$ : num [1:2] 0.865 0.135
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.775 0.225
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.384 0.616
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.146 0.854
#>   .. .. .. ..$ : num [1:2] 0.179 0.821
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.761 0.239
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.276 0.724
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.122 0.878
#>   .. .. .. ..$ : num [1:2] 0.162 0.838
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.719 0.281
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.273 0.727
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.101 0.899
#>   .. .. .. ..$ : num [1:2] 0.178 0.822
#>   .. .. ..$ :List of 9
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.455 0.545
#>   .. .. .. ..$ : num [1:2] 0.448 0.552
#>   .. .. .. ..$ : num [1:2] 0.148 0.852
#>   .. .. .. ..$ : num [1:2] 0.9496 0.0504
#>   .. .. .. ..$ : num [1:2] 0.397 0.603
#>   .. .. ..$ :List of 9
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.9545 0.0455
#>   .. .. .. ..$ : num [1:2] 0.384 0.616
#>   .. .. .. ..$ : num [1:2] 0.667 0.333
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.144 0.856
#>   .. .. .. ..$ : num [1:2] 0.205 0.795
#>   .. .. ..$ :List of 5
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.216 0.784
#>   .. .. .. ..$ : num [1:2] 0.9331 0.0669
#>   .. .. .. ..$ : num [1:2] 0.513 0.487
#>   .. .. ..$ :List of 9
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.399 0.601
#>   .. .. .. ..$ : num [1:2] 0.361 0.639
#>   .. .. .. ..$ : num [1:2] 0.136 0.864
#>   .. .. .. ..$ : num [1:2] 0.9683 0.0317
#>   .. .. .. ..$ : num [1:2] 0.9072 0.0928
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.769 0.231
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.299 0.701
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.152 0.848
#>   .. .. .. ..$ : num [1:2] 0.174 0.826
#>   .. .. ..$ :List of 5
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.199 0.801
#>   .. .. .. ..$ : num [1:2] 0.9091 0.0909
#>   .. .. .. ..$ : num [1:2] 0.46 0.54
#>   .. .. ..$ :List of 11
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.418 0.582
#>   .. .. .. ..$ : num [1:2] 0.327 0.673
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.9514 0.0486
#>   .. .. .. ..$ : num [1:2] 0.9348 0.0652
#>   .. .. .. ..$ : num [1:2] 0.146 0.854
#>   .. .. .. ..$ : num [1:2] 0.122 0.878
#>   .. .. ..$ :List of 5
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.209 0.791
#>   .. .. .. ..$ : num [1:2] 0.9388 0.0612
#>   .. .. .. ..$ : num [1:2] 0.5 0.5
#>   .. .. ..$ :List of 9
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.399 0.601
#>   .. .. .. ..$ : num [1:2] 0.34 0.66
#>   .. .. .. ..$ : num [1:2] 0.13 0.87
#>   .. .. .. ..$ : num [1:2] 0.9688 0.0312
#>   .. .. .. ..$ : num [1:2] 0.9783 0.0217
#>   .. .. ..$ :List of 9
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.686 0.314
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.898 0.102
#>   .. .. .. ..$ : num [1:2] 0.0955 0.9045
#>   .. .. .. ..$ : num [1:2] 0.448 0.552
#>   .. .. .. ..$ : num [1:2] 0.195 0.805
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.9123 0.0877
#>   .. .. .. ..$ : num [1:2] 0.419 0.581
#>   .. .. .. ..$ : num [1:2] 0.378 0.622
#>   .. .. .. ..$ : num [1:2] 0.168 0.832
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.233 0.767
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.268 0.732
#>   .. .. .. ..$ : num [1:2] 0.9677 0.0323
#>   .. .. .. ..$ : num [1:2] 0.889 0.111
#>   .. .. ..$ :List of 9
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.251 0.749
#>   .. .. .. ..$ : num [1:2] 0.468 0.532
#>   .. .. .. ..$ : num [1:2] 0.16 0.84
#>   .. .. .. ..$ : num [1:2] 0.9852 0.0148
#>   .. .. .. ..$ : num [1:2] 0.882 0.118
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.757 0.243
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.412 0.588
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.161 0.839
#>   .. .. .. ..$ : num [1:2] 0.171 0.829
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.671 0.329
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.242 0.758
#>   .. .. .. ..$ : num [1:2] 0.866 0.134
#>   .. .. .. ..$ : num [1:2] 0.144 0.856
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.716 0.284
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.341 0.659
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.107 0.893
#>   .. .. .. ..$ : num [1:2] 0.191 0.809
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.706 0.294
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.401 0.599
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.159 0.841
#>   .. .. .. ..$ : num [1:2] 0.185 0.815
#>   .. .. ..$ :List of 5
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.194 0.806
#>   .. .. .. ..$ : num [1:2] 0.9257 0.0743
#>   .. .. .. ..$ : num [1:2] 0.437 0.563
#>   .. .. ..$ :List of 9
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.457 0.543
#>   .. .. .. ..$ : num [1:2] 0.301 0.699
#>   .. .. .. ..$ : num [1:2] 0.196 0.804
#>   .. .. .. ..$ : num [1:2] 0.9416 0.0584
#>   .. .. .. ..$ : num [1:2] 0.891 0.109
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.629 0.371
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.678 0.322
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.143 0.857
#>   .. .. .. ..$ : num [1:2] 0.201 0.799
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.269 0.731
#>   .. .. .. ..$ : num [1:2] 0.9339 0.0661
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.379 0.621
#>   .. .. .. ..$ : num [1:2] 0.128 0.872
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.234 0.766
#>   .. .. .. ..$ : num [1:2] 0.607 0.393
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.891 0.109
#>   .. .. .. ..$ : num [1:2] 0.164 0.836
#>   .. .. ..$ :List of 11
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.494 0.506
#>   .. .. .. ..$ : num [1:2] 0.155 0.845
#>   .. .. .. ..$ : num [1:2] 0.9562 0.0438
#>   .. .. .. ..$ : num [1:2] 0.9245 0.0755
#>   .. .. .. ..$ : num [1:2] 0.35 0.65
#>   .. .. .. ..$ : num [1:2] 0.144 0.856
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.771 0.229
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.34 0.66
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.171 0.829
#>   .. .. .. ..$ : num [1:2] 0.165 0.835
#>   .. .. ..$ :List of 3
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.749 0.251
#>   .. .. .. ..$ : num [1:2] 0.175 0.825
#>   .. .. ..$ :List of 3
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.764 0.236
#>   .. .. .. ..$ : num [1:2] 0.219 0.781
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.204 0.796
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.491 0.509
#>   .. .. .. ..$ : num [1:2] 0.9652 0.0348
#>   .. .. .. ..$ : num [1:2] 0.9065 0.0935
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.189 0.811
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.427 0.573
#>   .. .. .. ..$ : num [1:2] 0.9784 0.0216
#>   .. .. .. ..$ : num [1:2] 0.9043 0.0957
#>   .. .. ..$ :List of 9
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.9375 0.0625
#>   .. .. .. ..$ : num [1:2] 0.491 0.509
#>   .. .. .. ..$ : num [1:2] 0.325 0.675
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.135 0.865
#>   .. .. .. ..$ : num [1:2] 0.172 0.828
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.24 0.76
#>   .. .. .. ..$ : num [1:2] 0.624 0.376
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.874 0.126
#>   .. .. .. ..$ : num [1:2] 0.142 0.858
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.9677 0.0323
#>   .. .. .. ..$ : num [1:2] 0.274 0.726
#>   .. .. .. ..$ : num [1:2] 0.431 0.569
#>   .. .. .. ..$ : num [1:2] 0.158 0.842
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.674 0.326
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.159 0.841
#>   .. .. .. ..$ : num [1:2] 0.9375 0.0625
#>   .. .. .. ..$ : num [1:2] 0.468 0.532
#>   .. .. ..$ :List of 11
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.9412 0.0588
#>   .. .. .. ..$ : num [1:2] 0.321 0.679
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.896 0.104
#>   .. .. .. ..$ : num [1:2] 0.586 0.414
#>   .. .. .. ..$ : num [1:2] 0.1 0.9
#>   .. .. .. ..$ : num [1:2] 0.147 0.853
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.89 0.11
#>   .. .. .. ..$ : num [1:2] 0.231 0.769
#>   .. .. .. ..$ : num [1:2] 0.434 0.566
#>   .. .. .. ..$ : num [1:2] 0.18 0.82
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.637 0.363
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.611 0.389
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.127 0.873
#>   .. .. .. ..$ : num [1:2] 0.263 0.737
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.976 0.024
#>   .. .. .. ..$ : num [1:2] 0.419 0.581
#>   .. .. .. ..$ : num [1:2] 0.455 0.545
#>   .. .. .. ..$ : num [1:2] 0.282 0.718
#>   .. .. ..$ :List of 9
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.9111 0.0889
#>   .. .. .. ..$ : num [1:2] 0.455 0.545
#>   .. .. .. ..$ : num [1:2] 0.379 0.621
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.137 0.863
#>   .. .. .. ..$ : num [1:2] 0.152 0.848
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.608 0.392
#>   .. .. .. ..$ : num [1:2] 0.463 0.537
#>   .. .. .. ..$ : num [1:2] 0.479 0.521
#>   .. .. .. ..$ : num [1:2] 0.16 0.84
#>   .. .. ..$ :List of 5
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.185 0.815
#>   .. .. .. ..$ : num [1:2] 0.9375 0.0625
#>   .. .. .. ..$ : num [1:2] 0.459 0.541
#>   .. .. ..$ :List of 5
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.198 0.802
#>   .. .. .. ..$ : num [1:2] 0.9289 0.0711
#>   .. .. .. ..$ : num [1:2] 0.486 0.514
#>   .. .. ..$ :List of 9
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.615 0.385
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.442 0.558
#>   .. .. .. ..$ : num [1:2] 0.177 0.823
#>   .. .. .. ..$ : num [1:2] 0.9375 0.0625
#>   .. .. .. ..$ : num [1:2] 0.141 0.859
#>   .. .. ..$ :List of 3
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.744 0.256
#>   .. .. .. ..$ : num [1:2] 0.199 0.801
#>   .. .. ..$ :List of 3
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.729 0.271
#>   .. .. .. ..$ : num [1:2] 0.218 0.782
#>   .. .. ..$ :List of 3
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.79 0.21
#>   .. .. .. ..$ : num [1:2] 0.21 0.79
#>   .. .. ..$ :List of 9
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.466 0.534
#>   .. .. .. ..$ : num [1:2] 0.459 0.541
#>   .. .. .. ..$ : num [1:2] 0.175 0.825
#>   .. .. .. ..$ : num [1:2] 0.9773 0.0227
#>   .. .. .. ..$ : num [1:2] 0.327 0.673
#>   .. .. ..$ :List of 9
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.493 0.507
#>   .. .. .. ..$ : num [1:2] 0.284 0.716
#>   .. .. .. ..$ : num [1:2] 0.148 0.852
#>   .. .. .. ..$ : num [1:2] 0.9664 0.0336
#>   .. .. .. ..$ : num [1:2] 0.833 0.167
#>   .. .. ..$ :List of 3
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.757 0.243
#>   .. .. .. ..$ : num [1:2] 0.211 0.789
#>   .. .. ..$ :List of 5
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.621 0.379
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.647 0.353
#>   .. .. .. ..$ : num [1:2] 0.187 0.813
#>   .. .. ..$ :List of 3
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.728 0.272
#>   .. .. .. ..$ : num [1:2] 0.214 0.786
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.204 0.796
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.447 0.553
#>   .. .. .. ..$ : num [1:2] 0.9718 0.0282
#>   .. .. .. ..$ : num [1:2] 0.869 0.131
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.9528 0.0472
#>   .. .. .. ..$ : num [1:2] 0.483 0.517
#>   .. .. .. ..$ : num [1:2] 0.269 0.731
#>   .. .. .. ..$ : num [1:2] 0.164 0.836
#>   .. .. ..$ :List of 11
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.449 0.551
#>   .. .. .. ..$ : num [1:2] 0.173 0.827
#>   .. .. .. ..$ : num [1:2] 0.9507 0.0493
#>   .. .. .. ..$ : num [1:2] 0.896 0.104
#>   .. .. .. ..$ : num [1:2] 0.31 0.69
#>   .. .. .. ..$ : num [1:2] 0.17 0.83
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.9076 0.0924
#>   .. .. .. ..$ : num [1:2] 0.241 0.759
#>   .. .. .. ..$ : num [1:2] 0.496 0.504
#>   .. .. .. ..$ : num [1:2] 0.185 0.815
#>   .. .. ..$ :List of 5
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.777 0.223
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.333 0.667
#>   .. .. .. ..$ : num [1:2] 0.173 0.827
#>   .. .. ..$ :List of 9
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.254 0.746
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.9333 0.0667
#>   .. .. .. ..$ : num [1:2] 0.899 0.101
#>   .. .. .. ..$ : num [1:2] 0.325 0.675
#>   .. .. .. ..$ : num [1:2] 0.172 0.828
#>   .. .. ..$ :List of 9
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.225 0.775
#>   .. .. .. ..$ : num [1:2] 0.496 0.504
#>   .. .. .. ..$ : num [1:2] 0.182 0.818
#>   .. .. .. ..$ : num [1:2] 0.9457 0.0543
#>   .. .. .. ..$ : num [1:2] 0.899 0.101
#>   .. .. ..$ :List of 9
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.9114 0.0886
#>   .. .. .. ..$ : num [1:2] 0.513 0.487
#>   .. .. .. ..$ : num [1:2] 0.39 0.61
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.151 0.849
#>   .. .. .. ..$ : num [1:2] 0.137 0.863
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.663 0.337
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.176 0.824
#>   .. .. .. ..$ : num [1:2] 0.881 0.119
#>   .. .. .. ..$ : num [1:2] 0.47 0.53
#>   .. .. ..$ :List of 5
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.223 0.777
#>   .. .. .. ..$ : num [1:2] 0.9274 0.0726
#>   .. .. .. ..$ : num [1:2] 0.523 0.477
#>   .. .. ..$ :List of 3
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.792 0.208
#>   .. .. .. ..$ : num [1:2] 0.217 0.783
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.224 0.776
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.515 0.485
#>   .. .. .. ..$ : num [1:2] 0.9357 0.0643
#>   .. .. .. ..$ : num [1:2] 0.9175 0.0825
#>   .. .. ..$ :List of 11
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.483 0.517
#>   .. .. .. ..$ : num [1:2] 0.268 0.732
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.963 0.037
#>   .. .. .. ..$ : num [1:2] 0.815 0.185
#>   .. .. .. ..$ : num [1:2] 0.182 0.818
#>   .. .. .. ..$ : num [1:2] 0.183 0.817
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.752 0.248
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.445 0.555
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.158 0.842
#>   .. .. .. ..$ : num [1:2] 0.177 0.823
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.246 0.754
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.483 0.517
#>   .. .. .. ..$ : num [1:2] 0.9574 0.0426
#>   .. .. .. ..$ : num [1:2] 0.838 0.162
#>   .. .. ..$ :List of 9
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.689 0.311
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.896 0.104
#>   .. .. .. ..$ : num [1:2] 0.529 0.471
#>   .. .. .. ..$ : num [1:2] 0.154 0.846
#>   .. .. .. ..$ : num [1:2] 0.174 0.826
#>   .. .. ..$ :List of 5
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.196 0.804
#>   .. .. .. ..$ : num [1:2] 0.9513 0.0487
#>   .. .. .. ..$ : num [1:2] 0.544 0.456
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.9309 0.0691
#>   .. .. .. ..$ : num [1:2] 0.497 0.503
#>   .. .. .. ..$ : num [1:2] 0.401 0.599
#>   .. .. .. ..$ : num [1:2] 0.177 0.823
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.712 0.288
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.393 0.607
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.108 0.892
#>   .. .. .. ..$ : num [1:2] 0.155 0.845
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.642 0.358
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.262 0.738
#>   .. .. .. ..$ : num [1:2] 0.877 0.123
#>   .. .. .. ..$ : num [1:2] 0.135 0.865
#>   .. .. ..$ :List of 7
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.639 0.361
#>   .. .. .. ..$ : num [1:2] 0.469 0.531
#>   .. .. .. ..$ : num [1:2] 0.493 0.507
#>   .. .. .. ..$ : num [1:2] 0.185 0.815
#>   .. .. ..$ :List of 9
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.566 0.434
#>   .. .. .. ..$ : num [1:2] 0.357 0.643
#>   .. .. .. ..$ : num [1:2] 0.152 0.848
#>   .. .. .. ..$ : num [1:2] 0.9291 0.0709
#>   .. .. .. ..$ : num [1:2] 0.957 0.043
#>   .. .. ..$ :List of 3
#>   .. .. .. ..$ : num(0) 
#>   .. .. .. ..$ : num [1:2] 0.74 0.26
#>   .. .. .. ..$ : num [1:2] 0.218 0.782
#>   .. .. .. [list output truncated]
#>   .. ..$ levels                    : chr [1:2] "Yes" "No"
#>   .. ..$ independent.variable.names: chr [1:2] "pclass" "sex"
#>   .. ..$ treetype                  : chr "Probability estimation"
#>   .. ..- attr(*, "class")= chr "ranger.forest"
#>   ..$ splitrule                : chr "gini"
#>   ..$ treetype                 : chr "Probability estimation"
#>   ..$ call                     : language (function (formula = NULL, data = NULL, num.trees = 500, mtry = NULL, importance = "none",      write.forest = TR| __truncated__ ...
#>   ..$ importance.mode          : chr "permutation"
#>   ..$ num.samples              : int 1043
#>   ..$ replace                  : logi TRUE
#>   ..$ model                    :Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	1043 obs. of  3 variables:
#>   .. ..$ survived: Factor w/ 2 levels "Yes","No": 1 1 2 2 2 1 1 2 1 2 ...
#>   .. ..$ pclass  : Factor w/ 3 levels "1st","2nd","3rd": 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..$ sex     : Factor w/ 2 levels "female","male": 1 2 1 2 1 2 1 2 1 2 ...
#>   .. ..- attr(*, "description")= chr "## Titanic\n\nThis dataset describes the survival status of individual passengers on the Titanic. The titanic d"| __truncated__
#>   ..- attr(*, "class")= chr "ranger"
#>  $ extra_args     : list()
#>  $ probability    : logi TRUE
#>  $ rv             : Factor w/ 2 levels "Yes","No": 1 1 2 2 2 1 1 2 1 2 ...
#>  $ not_vary       : chr(0) 
#>  $ df_name        : chr "titanic"
#>  $ vars           : chr [1:2] "pclass" "sex"
#>  $ rvar           : chr "survived"
#>  $ evar           : chr [1:2] "pclass" "sex"
#>  $ type           : chr "classification"
#>  $ lev            : chr "Yes"
#>  $ mtry           : NULL
#>  $ num.trees      : num 100
#>  $ min.node.size  : num 1
#>  $ sample.fraction: num 1
#>  $ replace        : logi TRUE
#>  $ wts            : NULL
#>  $ seed           : chr NA
#>  $ check          : chr ""
#>  $ data_filter    : chr ""
#>  - attr(*, "class")= chr [1:3] "rforest" "model" "list"

rforest(diamonds, "price", c("carat", "clarity"), type = "regression") %>% summary()

#> Random Forest (Ranger)
#> Type                 : Regression
#> Data                 : diamonds
#> Response variable    : price
#> Explanatory variables: carat, clarity 
#> Mtry                 : 
#> Number of trees      : 100 
#> Min node size        : 1 
#> Sample fraction      : 1 
#> Nr obs               : 3,000 
#> R-squared            : 0.919 
#> OOB prediction error : 1,125.346