scale_df {radiant.model}	R Documentation

Center or standardize variables in a data frame

Description

Center or standardize variables in a data frame

Usage

scale_df(dataset, center = TRUE, scale = TRUE, sf = 2, wts = NULL,
  calc = TRUE)

Arguments

dataset
Data frame

center
Center data (TRUE or FALSE)

scale
Scale data (TRUE or FALSE)

sf
Scaling factor (default is 2)

wts
Weights to use (default is NULL for no weights)

calc
Calculate mean and sd or use attributes attached to dat

Value

Scaled data frame

See Also

copy_attr to copy attributes from a training to a test dataset

[Package radiant.model version 1.0.0 Index]