test_specs {radiant.model}	R Documentation

Add interaction terms to list of test variables if needed

Description

Add interaction terms to list of test variables if needed

Usage

test_specs(tv, int)

Arguments

tv
List of variables to use for testing for regress or logistic

int
Interaction terms specified

Details

See https://radiant-rstats.github.io/docs/model/regress.html for an example in Radiant

Value

A vector of variables names to test

Examples

test_specs("a", "a:b")
test_specs("a", c("a:b", "b:c"))
test_specs("a", c("a:b", "b:c", "I(c^2)"))
test_specs(c("a", "b", "c"), c("a:b", "b:c", "I(c^2)"))

[Package radiant.model version 1.0.0 Index]