var_check {radiant.model}	R Documentation

Check if main effects for all interaction effects are included in the model

Description

Check if main effects for all interaction effects are included in the model

Usage

var_check(ev, cn, intv = c())

Arguments

ev
List of explanatory variables provided to regress or logistic

cn
Column names for all explanatory variables in the dataset

intv
Interaction terms specified

Details

If ':' is used to select a range evar is updated. See https://radiant-rstats.github.io/docs/model/regress.html for an example in Radiant

Value

vars is a vector of right-hand side variables, possibly with interactions, iv is the list of explanatory variables, and intv are interaction terms

Examples

var_check("a:d", c("a","b","c","d"))
var_check(c("a", "b"), c("a", "b"), "a:c")
var_check(c("a", "b"), c("a", "b"), "a:c")
var_check(c("a", "b"), c("a", "b"), c("a:c", "I(b^2)"))

[Package radiant.model version 1.0.0 Index]