Write coefficient table for linear and logistic regression

Write coefficient table for linear and logistic regression

write.coeff(object, file = "", sort = FALSE, intercept = TRUE)

Arguments

object
A fitted model object of class regress or logistic

file
A character string naming a file. "" indicates output to the console

sort
Sort table by variable importance

intercept
Include the intercept in the output (TRUE or FALSE). TRUE is the default

Details

Write coefficients and importance scores to csv or or return as a data.frame

Examples

regress(diamonds, rvar = "price", evar = c("carat", "clarity", "x"), check = "standardize") %>%
  write.coeff(sort = TRUE) %>%
  format_df(dec = 3)

#> Standardized coefficients selected
#> 

#>           label coefficient std.error t.value p.value sig_star dummy  mean
#> 1   (Intercept)      -0.506     0.024 -20.833   0.000      *** 0.000 1.000
#> 2         carat       1.283     0.027  47.278   0.000      *** 0.000 0.794
#> 3    clarity|IF       0.653     0.029  22.498   0.000      *** 1.000 0.033
#> 4  clarity|VVS2       0.640     0.026  24.495   0.000      *** 1.000 0.095
#> 5  clarity|VVS1       0.629     0.027  23.627   0.000      *** 1.000 0.075
#> 6   clarity|VS1       0.567     0.025  22.308   0.000      *** 1.000 0.147
#> 7   clarity|VS2       0.541     0.025  21.590   0.000      *** 1.000 0.220
#> 8   clarity|SI1       0.462     0.025  18.566   0.000      *** 1.000 0.240
#> 9   clarity|SI2       0.359     0.025  14.334   0.000      *** 1.000 0.176
#> 10            x      -0.282     0.027 -10.300   0.000      *** 0.000 5.722
#>       sd   min   max importance
#> 1  0.000 1.000 1.000      0.000
#> 2  0.474 0.200 3.000      1.283
#> 3  0.179 0.000 1.000      0.653
#> 4  0.293 0.000 1.000      0.640
#> 5  0.263 0.000 1.000      0.629
#> 6  0.354 0.000 1.000      0.567
#> 7  0.415 0.000 1.000      0.541
#> 8  0.427 0.000 1.000      0.462
#> 9  0.381 0.000 1.000      0.359
#> 10 1.124 3.730 9.420      0.282