Gradient Boosted Trees using XGBoost

Gradient Boosted Trees using XGBoost

gbt(dataset, rvar, evar, type = "classification", lev = "",
  max_depth = 6, learning_rate = 0.3, min_split_loss = 0,
  min_child_weight = 1, subsample = 1, nrounds = 100,
  early_stopping_rounds = 10, nthread = 12, wts = "None",
  seed = NA, data_filter = "", envir = parent.frame(), ...)

Arguments

dataset
Dataset

rvar
The response variable in the model

evar
Explanatory variables in the model

type
Model type (i.e., "classification" or "regression")

lev
Level to use as the first column in prediction output

max_depth
Maximum 'depth' of tree

learning_rate
Learning rate (eta)

min_split_loss
Minimal improvement (gamma)

min_child_weight
Minimum number of instances allowed in each node

subsample
Subsample ratio of the training instances (0-1)

nrounds
Number of trees to create

early_stopping_rounds
Early stopping rule

nthread
Number of parallel threads to use. Defaults to 12 if available

wts
Weights to use in estimation

seed
Random seed to use as the starting point

data_filter
Expression entered in, e.g., Data > View to filter the dataset in Radiant. The expression should be a string (e.g., "price > 10000")

envir
Environment to extract data from

...
Further arguments to pass to xgboost

Value

A list with all variables defined in gbt as an object of class gbt

Details

See https://radiant-rstats.github.io/docs/model/gbt.html for an example in Radiant

See also

Examples

gbt(titanic, "survived", c("pclass", "sex"), lev = "Yes") %>% summary()

#> Gradient Boosted Trees (XGBoost)
#> Type                 : Classification
#> Data                 : titanic
#> Response variable    : survived
#> Level                : Yes in survived
#> Explanatory variables: pclass, sex 
#> Max depth            : 6 
#> Learning rate (eta)  : 0.3 
#> Min split loss       : 0 
#> Min child weight     : 1 
#> Sub-sample           : 1 
#> Nr of rounds (trees) : 100 
#> Early stopping rounds: 10 
#> Nr obs               : 1,043 
#> 
#> Iteration history:
#> 
#> [1]	train-auc:0.824504 
#> [2]	train-auc:0.824504 
#> [3]	train-auc:0.826267 
#> [4]	train-auc:0.826267 
#> [5]	train-auc:0.826267 
#> [6]	train-auc:0.826267 
#> [7]	train-auc:0.826267 
#> [8]	train-auc:0.826267 
#> [9]	train-auc:0.826267 
#> [10]	train-auc:0.826267 
#> [11]	train-auc:0.826267 
#> [12]	train-auc:0.826267 
#> [13]	train-auc:0.826267 
#> Stopping. Best iteration:
#> [3]	train-auc:0.826267

gbt(titanic, "survived", c("pclass", "sex")) %>% str()

#> List of 26
#>  $ check                : chr ""
#>  $ model                :List of 14
#>   ..$ handle         :Class 'xgb.Booster.handle' <externalptr> 
#>   ..$ raw            : raw [1:7517] 00 00 00 80 ...
#>   ..$ best_iteration : int 3
#>   ..$ best_ntreelimit: num 3
#>   ..$ best_score     : Named num 0.826
#>   .. ..- attr(*, "names")= chr "train-auc"
#>   ..$ niter          : int 13
#>   ..$ evaluation_log :Classes ‘data.table’ and 'data.frame':	13 obs. of  2 variables:
#>   .. ..$ iter     : num [1:13] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. ..$ train_auc: num [1:13] 0.825 0.825 0.826 0.826 0.826 ...
#>   .. ..- attr(*, ".internal.selfref")=<externalptr> 
#>   ..$ call           : language xgb.train(params = params, data = dtrain, nrounds = nrounds, watchlist = watchlist,      verbose = verbose, print| __truncated__ ...
#>   ..$ params         :List of 9
#>   .. ..$ max_depth       : num 6
#>   .. ..$ learning_rate   : num 0.3
#>   .. ..$ min_split_loss  : num 0
#>   .. ..$ min_child_weight: num 1
#>   .. ..$ subsample       : num 1
#>   .. ..$ nthread         : num 12
#>   .. ..$ objective       : chr "binary:logistic"
#>   .. ..$ eval_metric     : chr "auc"
#>   .. ..$ silent          : num 1
#>   ..$ callbacks      :List of 3
#>   .. ..$ cb.print.evaluation:function (env = parent.frame())  
#>   .. .. ..- attr(*, "call")= language cb.print.evaluation(period = print_every_n)
#>   .. .. ..- attr(*, "name")= chr "cb.print.evaluation"
#>   .. ..$ cb.evaluation.log  :function (env = parent.frame(), finalize = FALSE)  
#>   .. .. ..- attr(*, "call")= language cb.evaluation.log()
#>   .. .. ..- attr(*, "name")= chr "cb.evaluation.log"
#>   .. ..$ cb.early.stop      :function (env = parent.frame(), finalize = FALSE)  
#>   .. .. ..- attr(*, "call")= language cb.early.stop(stopping_rounds = early_stopping_rounds, maximize = maximize,      verbose = verbose)
#>   .. .. ..- attr(*, "name")= chr "cb.early.stop"
#>   ..$ feature_names  : chr [1:3] "pclass2nd" "pclass3rd" "sexmale"
#>   ..$ nfeatures      : int 3
#>   ..$ importance     :Classes ‘data.table’ and 'data.frame':	3 obs. of  4 variables:
#>   .. ..$ Feature  : chr [1:3] "sexmale" "pclass3rd" "pclass2nd"
#>   .. ..$ Gain     : num [1:3] 0.7284 0.236 0.0356
#>   .. ..$ Cover    : num [1:3] 0.392 0.359 0.249
#>   .. ..$ Frequency: num [1:3] 0.206 0.413 0.381
#>   .. ..- attr(*, ".internal.selfref")=<externalptr> 
#>   ..$ model          :Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	1043 obs. of  3 variables:
#>   .. ..$ survived: Factor w/ 2 levels "Yes","No": 1 1 2 2 2 1 1 2 1 2 ...
#>   .. ..$ pclass  : Factor w/ 3 levels "1st","2nd","3rd": 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..$ sex     : Factor w/ 2 levels "female","male": 1 2 1 2 1 2 1 2 1 2 ...
#>   .. ..- attr(*, "description")= chr "## Titanic\n\nThis dataset describes the survival status of individual passengers on the Titanic. The titanic d"| __truncated__
#>   ..- attr(*, "class")= chr "xgb.Booster"
#>  $ output               : chr [1:18] "[1]\ttrain-auc:0.824504 " "Will train until train_auc hasn't improved in 10 rounds." "" "[2]\ttrain-auc:0.824504 " ...
#>  $ check_args           :function (arg, default, inp = gbt_input)  
#>  $ extra_args_names     : NULL
#>  $ extra_args           : list()
#>  $ gbt_input            :List of 10
#>   ..$ max_depth            : num 6
#>   ..$ learning_rate        : num 0.3
#>   ..$ min_split_loss       : num 0
#>   ..$ nrounds              : num 100
#>   ..$ min_child_weight     : num 1
#>   ..$ subsample            : num 1
#>   ..$ early_stopping_rounds: num 10
#>   ..$ nthread              : num 12
#>   ..$ objective            : chr "binary:logistic"
#>   ..$ eval_metric          : chr "auc"
#>  $ not_vary             : chr(0) 
#>  $ nr_obs               : int 1043
#>  $ df_name              : chr "titanic"
#>  $ vars                 : chr [1:2] "pclass" "sex"
#>  $ rvar                 : chr "survived"
#>  $ evar                 : chr [1:2] "pclass" "sex"
#>  $ type                 : chr "classification"
#>  $ lev                  : chr "Yes"
#>  $ max_depth            : num 6
#>  $ learning_rate        : num 0.3
#>  $ min_split_loss       : num 0
#>  $ min_child_weight     : num 1
#>  $ subsample            : num 1
#>  $ nrounds              : num 100
#>  $ early_stopping_rounds: num 10
#>  $ nthread              : num 12
#>  $ wts                  : NULL
#>  $ seed                 : chr NA
#>  $ data_filter          : chr ""
#>  - attr(*, "class")= chr [1:3] "gbt" "model" "list"

gbt(titanic, "survived", c("pclass", "sex"), eval_metric = paste0("error@", 0.5/6)) %>% str()

#> List of 26
#>  $ check                : chr ""
#>  $ model                :List of 14
#>   ..$ handle         :Class 'xgb.Booster.handle' <externalptr> 
#>   ..$ raw            : raw [1:6445] 00 00 00 80 ...
#>   ..$ best_iteration : int 1
#>   ..$ best_ntreelimit: num 1
#>   ..$ best_score     : Named num 0.593
#>   .. ..- attr(*, "names")= chr "train-error@0.0833333"
#>   ..$ niter          : int 11
#>   ..$ evaluation_log :Classes ‘data.table’ and 'data.frame':	11 obs. of  2 variables:
#>   .. ..$ iter                 : num [1:11] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. ..$ train_error@0.0833333: num [1:11] 0.593 0.593 0.593 0.593 0.593 ...
#>   .. ..- attr(*, ".internal.selfref")=<externalptr> 
#>   ..$ call           : language xgb.train(params = params, data = dtrain, nrounds = nrounds, watchlist = watchlist,      verbose = verbose, print| __truncated__ ...
#>   ..$ params         :List of 9
#>   .. ..$ max_depth       : num 6
#>   .. ..$ learning_rate   : num 0.3
#>   .. ..$ min_split_loss  : num 0
#>   .. ..$ min_child_weight: num 1
#>   .. ..$ subsample       : num 1
#>   .. ..$ nthread         : num 12
#>   .. ..$ objective       : chr "binary:logistic"
#>   .. ..$ eval_metric     : chr "error@0.0833333333333333"
#>   .. ..$ silent          : num 1
#>   ..$ callbacks      :List of 3
#>   .. ..$ cb.print.evaluation:function (env = parent.frame())  
#>   .. .. ..- attr(*, "call")= language cb.print.evaluation(period = print_every_n)
#>   .. .. ..- attr(*, "name")= chr "cb.print.evaluation"
#>   .. ..$ cb.evaluation.log  :function (env = parent.frame(), finalize = FALSE)  
#>   .. .. ..- attr(*, "call")= language cb.evaluation.log()
#>   .. .. ..- attr(*, "name")= chr "cb.evaluation.log"
#>   .. ..$ cb.early.stop      :function (env = parent.frame(), finalize = FALSE)  
#>   .. .. ..- attr(*, "call")= language cb.early.stop(stopping_rounds = early_stopping_rounds, maximize = maximize,      verbose = verbose)
#>   .. .. ..- attr(*, "name")= chr "cb.early.stop"
#>   ..$ feature_names  : chr [1:3] "pclass2nd" "pclass3rd" "sexmale"
#>   ..$ nfeatures      : int 3
#>   ..$ importance     :Classes ‘data.table’ and 'data.frame':	3 obs. of  4 variables:
#>   .. ..$ Feature  : chr [1:3] "sexmale" "pclass3rd" "pclass2nd"
#>   .. ..$ Gain     : num [1:3] 0.7291 0.2356 0.0353
#>   .. ..$ Cover    : num [1:3] 0.393 0.364 0.243
#>   .. ..$ Frequency: num [1:3] 0.208 0.415 0.377
#>   .. ..- attr(*, ".internal.selfref")=<externalptr> 
#>   ..$ model          :Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	1043 obs. of  3 variables:
#>   .. ..$ survived: Factor w/ 2 levels "Yes","No": 1 1 2 2 2 1 1 2 1 2 ...
#>   .. ..$ pclass  : Factor w/ 3 levels "1st","2nd","3rd": 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..$ sex     : Factor w/ 2 levels "female","male": 1 2 1 2 1 2 1 2 1 2 ...
#>   .. ..- attr(*, "description")= chr "## Titanic\n\nThis dataset describes the survival status of individual passengers on the Titanic. The titanic d"| __truncated__
#>   ..- attr(*, "class")= chr "xgb.Booster"
#>  $ output               : chr [1:16] "[1]\ttrain-error@0.0833333:0.592522 " "Will train until train_error@0.0833333 hasn't improved in 10 rounds." "" "[2]\ttrain-error@0.0833333:0.592522 " ...
#>  $ check_args           :function (arg, default, inp = gbt_input)  
#>  $ extra_args_names     : chr "eval_metric"
#>  $ extra_args           :List of 1
#>   ..$ eval_metric: chr "error@0.0833333333333333"
#>  $ gbt_input            :List of 10
#>   ..$ max_depth            : num 6
#>   ..$ learning_rate        : num 0.3
#>   ..$ min_split_loss       : num 0
#>   ..$ nrounds              : num 100
#>   ..$ min_child_weight     : num 1
#>   ..$ subsample            : num 1
#>   ..$ early_stopping_rounds: num 10
#>   ..$ nthread              : num 12
#>   ..$ objective            : chr "binary:logistic"
#>   ..$ eval_metric          : chr "error@0.0833333333333333"
#>  $ not_vary             : chr(0) 
#>  $ nr_obs               : int 1043
#>  $ df_name              : chr "titanic"
#>  $ vars                 : chr [1:2] "pclass" "sex"
#>  $ rvar                 : chr "survived"
#>  $ evar                 : chr [1:2] "pclass" "sex"
#>  $ type                 : chr "classification"
#>  $ lev                  : chr "Yes"
#>  $ max_depth            : num 6
#>  $ learning_rate        : num 0.3
#>  $ min_split_loss       : num 0
#>  $ min_child_weight     : num 1
#>  $ subsample            : num 1
#>  $ nrounds              : num 100
#>  $ early_stopping_rounds: num 10
#>  $ nthread              : num 12
#>  $ wts                  : NULL
#>  $ seed                 : chr NA
#>  $ data_filter          : chr ""
#>  - attr(*, "class")= chr [1:3] "gbt" "model" "list"

gbt(diamonds, "price", c("carat", "clarity"), type = "regression") %>% summary()

#> Gradient Boosted Trees (XGBoost)
#> Type                 : Regression
#> Data                 : diamonds
#> Response variable    : price
#> Explanatory variables: carat, clarity 
#> Max depth            : 6 
#> Learning rate (eta)  : 0.3 
#> Min split loss       : 0 
#> Min child weight     : 1 
#> Sub-sample           : 1 
#> Nr of rounds (trees) : 100 
#> Early stopping rounds: 10 
#> Nr obs               : 3,000 
#> 
#> Iteration history:
#> 
#> [1]	train-rmse:4004.124512 
#> [2]	train-rmse:2934.489990 
#> [3]	train-rmse:2209.099365 
#> [4]	train-rmse:1731.832275 
#> [5]	train-rmse:1429.542236 
#> [6]	train-rmse:1236.558594 
#> [7]	train-rmse:1112.986206 
#> [8]	train-rmse:1041.411133 
#> [9]	train-rmse:994.494568 
#> [10]	train-rmse:968.976013 
#> ...
#> [91]	train-rmse:747.640930 
#> [92]	train-rmse:747.167053 
#> [93]	train-rmse:745.880615 
#> [94]	train-rmse:744.933716 
#> [95]	train-rmse:744.060425 
#> [96]	train-rmse:743.978699 
#> [97]	train-rmse:743.872742 
#> [98]	train-rmse:743.621277 
#> [99]	train-rmse:743.458923 
#> [100]	train-rmse:743.281738 

rig_wrap <- function(preds, dtrain) {
  labels <- xgboost::getinfo(dtrain, "label")
  value <- rig(preds, labels, lev = 1)
  list(metric = "rig", value = value)
}
gbt(titanic, "survived", c("pclass", "sex"), eval_metric = rig_wrap, maximize = TRUE) %>% str()

#> List of 26
#>  $ check                : chr ""
#>  $ model                :List of 14
#>   ..$ handle         :Class 'xgb.Booster.handle' <externalptr> 
#>   ..$ raw            : raw [1:17685] 00 00 00 80 ...
#>   ..$ best_iteration : int 25
#>   ..$ best_ntreelimit: num 25
#>   ..$ best_score     : Named num 0.315
#>   .. ..- attr(*, "names")= chr "train-rig"
#>   ..$ niter          : int 35
#>   ..$ evaluation_log :Classes ‘data.table’ and 'data.frame':	35 obs. of  2 variables:
#>   .. ..$ iter     : num [1:35] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. ..$ train_rig: num [1:35] 0.124 0.204 0.249 0.276 0.292 ...
#>   .. ..- attr(*, ".internal.selfref")=<externalptr> 
#>   ..$ call           : language xgb.train(params = params, data = dtrain, nrounds = nrounds, watchlist = watchlist,      verbose = verbose, print| __truncated__ ...
#>   ..$ params         :List of 8
#>   .. ..$ max_depth       : num 6
#>   .. ..$ learning_rate   : num 0.3
#>   .. ..$ min_split_loss  : num 0
#>   .. ..$ min_child_weight: num 1
#>   .. ..$ subsample       : num 1
#>   .. ..$ nthread         : num 12
#>   .. ..$ objective       : chr "binary:logistic"
#>   .. ..$ silent          : num 1
#>   ..$ callbacks      :List of 3
#>   .. ..$ cb.print.evaluation:function (env = parent.frame())  
#>   .. .. ..- attr(*, "call")= language cb.print.evaluation(period = print_every_n)
#>   .. .. ..- attr(*, "name")= chr "cb.print.evaluation"
#>   .. ..$ cb.evaluation.log  :function (env = parent.frame(), finalize = FALSE)  
#>   .. .. ..- attr(*, "call")= language cb.evaluation.log()
#>   .. .. ..- attr(*, "name")= chr "cb.evaluation.log"
#>   .. ..$ cb.early.stop      :function (env = parent.frame(), finalize = FALSE)  
#>   .. .. ..- attr(*, "call")= language cb.early.stop(stopping_rounds = early_stopping_rounds, maximize = maximize,      verbose = verbose)
#>   .. .. ..- attr(*, "name")= chr "cb.early.stop"
#>   ..$ feature_names  : chr [1:3] "pclass2nd" "pclass3rd" "sexmale"
#>   ..$ nfeatures      : int 3
#>   ..$ importance     :Classes ‘data.table’ and 'data.frame':	3 obs. of  4 variables:
#>   .. ..$ Feature  : chr [1:3] "sexmale" "pclass3rd" "pclass2nd"
#>   .. ..$ Gain     : num [1:3] 0.7281 0.2362 0.0357
#>   .. ..$ Cover    : num [1:3] 0.415 0.333 0.252
#>   .. ..$ Frequency: num [1:3] 0.218 0.395 0.388
#>   .. ..- attr(*, ".internal.selfref")=<externalptr> 
#>   ..$ model          :Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	1043 obs. of  3 variables:
#>   .. ..$ survived: Factor w/ 2 levels "Yes","No": 1 1 2 2 2 1 1 2 1 2 ...
#>   .. ..$ pclass  : Factor w/ 3 levels "1st","2nd","3rd": 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..$ sex     : Factor w/ 2 levels "female","male": 1 2 1 2 1 2 1 2 1 2 ...
#>   .. ..- attr(*, "description")= chr "## Titanic\n\nThis dataset describes the survival status of individual passengers on the Titanic. The titanic d"| __truncated__
#>   ..- attr(*, "class")= chr "xgb.Booster"
#>  $ output               : chr [1:40] "[1]\ttrain-rig:0.124424 " "Will train until train_rig hasn't improved in 10 rounds." "" "[2]\ttrain-rig:0.203935 " ...
#>  $ check_args           :function (arg, default, inp = gbt_input)  
#>  $ extra_args_names     : chr [1:2] "eval_metric" "maximize"
#>  $ extra_args           :List of 2
#>   ..$ eval_metric:function (preds, dtrain)  
#>   .. ..- attr(*, "srcref")= 'srcref' int [1:8] 5 13 9 1 13 1 5 9
#>   .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0xf3cc3a0> 
#>   ..$ maximize   : logi TRUE
#>  $ gbt_input            :List of 11
#>   ..$ max_depth            : num 6
#>   ..$ learning_rate        : num 0.3
#>   ..$ min_split_loss       : num 0
#>   ..$ nrounds              : num 100
#>   ..$ min_child_weight     : num 1
#>   ..$ subsample            : num 1
#>   ..$ early_stopping_rounds: num 10
#>   ..$ nthread              : num 12
#>   ..$ objective            : chr "binary:logistic"
#>   ..$ eval_metric          :function (preds, dtrain)  
#>   .. ..- attr(*, "srcref")= 'srcref' int [1:8] 5 13 9 1 13 1 5 9
#>   .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0xf3cc3a0> 
#>   ..$ maximize             : logi TRUE
#>  $ not_vary             : chr(0) 
#>  $ nr_obs               : int 1043
#>  $ df_name              : chr "titanic"
#>  $ vars                 : chr [1:2] "pclass" "sex"
#>  $ rvar                 : chr "survived"
#>  $ evar                 : chr [1:2] "pclass" "sex"
#>  $ type                 : chr "classification"
#>  $ lev                  : chr "Yes"
#>  $ max_depth            : num 6
#>  $ learning_rate        : num 0.3
#>  $ min_split_loss       : num 0
#>  $ min_child_weight     : num 1
#>  $ subsample            : num 1
#>  $ nrounds              : num 100
#>  $ early_stopping_rounds: num 10
#>  $ nthread              : num 12
#>  $ wts                  : NULL
#>  $ seed                 : chr NA
#>  $ data_filter          : chr ""
#>  - attr(*, "class")= chr [1:3] "gbt" "model" "list"