logistic {radiant.model}	R Documentation

Logistic regression

Description

Logistic regression

Usage

logistic(dataset, rvar, evar, lev = "", int = "", wts = "None",
  check = "", ci_type, data_filter = "")

Arguments

dataset
Dataset

rvar
The response variable in the model

evar
Explanatory variables in the model

lev
The level in the response variable defined as _success_

int
Interaction term to include in the model

wts
Weights to use in estimation

check
Use "standardize" to see standardized coefficient estimates. Use "stepwise-backward" (or "stepwise-forward", or "stepwise-both") to apply step-wise selection of variables in estimation. Add "robust" for robust estimation of standard errors (HC1)

ci_type
To use the profile-likelihood (rather than Wald) for confidence intervals use "profile". For datasets with more than 5,000 rows the Wald method will be used, unless "profile" is explicitly set

data_filter
Expression entered in, e.g., Data > View to filter the dataset in Radiant. The expression should be a string (e.g., "price > 10000")

Details

See https://radiant-rstats.github.io/docs/model/logistic.html for an example in Radiant

Value

A list with all variables defined in logistic as an object of class logistic

See Also

summary.logistic to summarize the results

plot.logistic to plot the results

predict.logistic to generate predictions

plot.model.predict to plot prediction output

Examples

logistic(titanic, "survived", c("pclass", "sex"), lev = "Yes") %>% summary()
logistic(titanic, "survived", c("pclass", "sex")) %>% str()

[Package radiant.model version 1.0.0 Index]