nn {radiant.model}	R Documentation

Neural Networks

Description

Neural Networks

Usage

nn(dataset, rvar, evar, type = "classification", lev = "", size = 1,
  decay = 0.5, wts = "None", seed = NA, check = "standardize",
  data_filter = "")

Arguments

dataset
Dataset

rvar
The response variable in the model

evar
Explanatory variables in the model

type
Model type (i.e., "classification" or "regression")

lev
The level in the response variable defined as _success_

size
Number of units (nodes) in the hidden layer

decay
Parameter decay

wts
Weights to use in estimation

seed
Random seed to use as the starting point

check
Optional estimation parameters ("standardize" is the default)

data_filter
Expression entered in, e.g., Data > View to filter the dataset in Radiant. The expression should be a string (e.g., "price > 10000")

Details

See https://radiant-rstats.github.io/docs/model/nn.html for an example in Radiant

Value

A list with all variables defined in nn as an object of class nn

See Also

summary.nn to summarize results

plot.nn to plot results

predict.nn for prediction

Examples

nn(titanic, "survived", c("pclass", "sex"), lev = "Yes") %>% summary()
nn(titanic, "survived", c("pclass", "sex")) %>% str()
nn(diamonds, "price", c("carat", "clarity"), type = "regression") %>% summary()

[Package radiant.model version 1.0.0 Index]