write.coeff {radiant.model}	R Documentation

Write coefficient table for linear and logistic regression

Description

Write coefficient table for linear and logistic regression

Usage

write.coeff(object, file = "", sort = FALSE, intercept = TRUE)

Arguments

object
A fitted model object of class regress or logistic

file
A character string naming a file. "" indicates output to the console

sort
Sort table by variable importance

intercept
Include the intercept in the output (TRUE or FALSE). TRUE is the default

Details

Write coefficients and importance scores to csv or or return as a data.frame

Examples

regress(
 diamonds, rvar = "price", evar = c("carat", "clarity", "color", "x"),
 int = c("carat:clarity", "clarity:color", "I(x^2)"), check = "standardize"
) %>%
  write.coeff(sort = TRUE) %>%
  format_df(dec = 3)

[Package radiant.model version 1.0.0 Index]