Write coefficient table for linear and logistic regression

Write coefficient table for linear and logistic regression

write.coeff(object, file = "", sort = FALSE, intercept = TRUE)

Arguments

object
A fitted model object of class regress or logistic

file
A character string naming a file. "" indicates output to the console

sort
Sort table by variable importance

intercept
Include the intercept in the output (TRUE or FALSE). TRUE is the default

Details

Write coefficients and importance scores to csv or or return as a data.frame

Examples

regress(
 diamonds, rvar = "price", evar = c("carat", "clarity", "color", "x"),
 int = c("carat:clarity", "clarity:color", "I(x^2)"), check = "standardize"
) %>%
  write.coeff(sort = TRUE) %>%
  format_df(dec = 3)

#> Standardized coefficients selected
#> 

#>                   label coefficient std.error t.value p.value sig_star dummy
#> 1           (Intercept)      -0.162     0.059  -2.738   0.006      **      0
#> 2      carat:clarity|IF       1.089     0.061  17.729   0.000      ***     0
#> 3    carat:clarity|VVS1       1.020     0.052  19.677   0.000      ***     0
#> 4            clarity|IF       1.015     0.098  10.349   0.000      ***     1
#> 5    carat:clarity|VVS2       1.001     0.049  20.587   0.000      ***     0
#> 6    clarity|IF:color|J      -0.946     0.147  -6.425   0.000      ***     1
#> 7     carat:clarity|VS1       0.743     0.046  16.014   0.000      ***     0
#> 8     carat:clarity|VS2       0.705     0.046  15.430   0.000      ***     0
#> 9    clarity|IF:color|H      -0.610     0.108  -5.654   0.000      ***     1
#> 10   clarity|IF:color|I      -0.606     0.114  -5.298   0.000      ***     1
#> 11    carat:clarity|SI1       0.603     0.045  13.277   0.000      ***     0
#> 12   clarity|IF:color|F      -0.572     0.105  -5.456   0.000      ***     1
#> 13   clarity|IF:color|E      -0.534     0.114  -4.679   0.000      ***     1
#> 14    carat:clarity|SI2       0.481     0.045  10.682   0.000      ***     0
#> 15   clarity|IF:color|G      -0.441     0.115  -3.816   0.000      ***     1
#> 16         clarity|VVS2       0.349     0.062   5.639   0.000      ***     1
#> 17               I(x^2)       0.345     0.020  17.227   0.000      ***     0
#> 18         clarity|VVS1       0.339     0.066   5.159   0.000      ***     1
#> 19                carat       0.286     0.069   4.151   0.000      ***     0
#> 20 clarity|VVS2:color|J      -0.284     0.107  -2.646   0.008      **      1
#> 21          clarity|VS1       0.252     0.061   4.114   0.000      ***     1
#> 22  clarity|VS2:color|J      -0.232     0.099  -2.338   0.019      *       1
#> 23              color|G      -0.231     0.079  -2.912   0.004      **      1
#> 24  clarity|VS2:color|G       0.201     0.081   2.494   0.013      *       1
#> 25 clarity|VVS1:color|J      -0.192     0.120  -1.604   0.109              1
#> 26  clarity|VS1:color|J      -0.192     0.101  -1.908   0.057      .       1
#> 27 clarity|VVS1:color|G       0.190     0.085   2.239   0.025      *       1
#> 28          clarity|VS2       0.177     0.060   2.929   0.003      **      1
#> 29  clarity|SI2:color|G       0.176     0.081   2.169   0.030      *       1
#> 30  clarity|SI1:color|G       0.175     0.081   2.162   0.031      *       1
#> 31  clarity|VS1:color|G       0.155     0.081   1.902   0.057      .       1
#> 32              color|I      -0.144     0.075  -1.926   0.054      .       1
#> 33 clarity|VVS2:color|G       0.144     0.082   1.754   0.080      .       1
#> 34  clarity|SI1:color|J      -0.133     0.099  -1.352   0.176              1
#> 35              color|J      -0.121     0.097  -1.255   0.210              1
#> 36  clarity|SI2:color|J      -0.116     0.100  -1.161   0.246              1
#> 37 clarity|VVS2:color|I      -0.108     0.081  -1.330   0.184              1
#> 38          clarity|SI1       0.103     0.060   1.719   0.086      .       1
#> 39              color|H      -0.101     0.068  -1.487   0.137              1
#> 40  clarity|SI2:color|E       0.100     0.077   1.300   0.194              1
#> 41 clarity|VVS1:color|E       0.096     0.082   1.170   0.242              1
#> 42  clarity|VS1:color|I      -0.095     0.078  -1.214   0.225              1
#> 43              color|E      -0.087     0.075  -1.161   0.246              1
#> 44  clarity|VS2:color|E       0.079     0.077   1.037   0.300              1
#> 45  clarity|SI1:color|E       0.073     0.076   0.961   0.337              1
#> 46  clarity|VS1:color|E       0.066     0.078   0.846   0.397              1
#> 47 clarity|VVS2:color|E       0.063     0.079   0.799   0.425              1
#> 48  clarity|VS1:color|H      -0.059     0.071  -0.823   0.410              1
#> 49 clarity|VVS1:color|F       0.047     0.072   0.647   0.518              1
#> 50 clarity|VVS1:color|I      -0.046     0.084  -0.553   0.580              1
#> 51  clarity|VS2:color|I      -0.038     0.077  -0.486   0.627              1
#> 52  clarity|VS1:color|F      -0.033     0.068  -0.489   0.625              1
#> 53              color|F      -0.032     0.065  -0.497   0.619              1
#> 54  clarity|SI1:color|H       0.029     0.070   0.420   0.675              1
#> 55 clarity|VVS1:color|H       0.026     0.076   0.348   0.728              1
#> 56          clarity|SI2       0.025     0.061   0.406   0.685              1
#> 57  clarity|VS2:color|F       0.024     0.067   0.362   0.718              1
#> 58  clarity|VS2:color|H      -0.020     0.070  -0.281   0.779              1
#> 59  clarity|SI1:color|I      -0.020     0.077  -0.257   0.797              1
#> 60  clarity|SI2:color|F       0.018     0.067   0.270   0.787              1
#> 61 clarity|VVS2:color|H      -0.016     0.073  -0.222   0.824              1
#> 62  clarity|SI2:color|H       0.014     0.071   0.196   0.844              1
#> 63 clarity|VVS2:color|F       0.012     0.069   0.175   0.861              1
#> 64  clarity|SI2:color|I      -0.012     0.078  -0.151   0.880              1
#> 65  clarity|SI1:color|F       0.010     0.067   0.157   0.875              1
#> 66                    x       0.009     0.047   0.197   0.844              0
#>      mean     sd    min    max importance
#> 1   1.000  0.000  1.000  1.000      0.000
#> 2   0.015  0.097  0.000  1.530      1.089
#> 3   0.039  0.158  0.000  1.580      1.020
#> 4   0.033  0.179  0.000  1.000      1.015
#> 5   0.056  0.207  0.000  1.710      1.001
#> 6   0.001  0.026  0.000  1.000      0.946
#> 7   0.107  0.303  0.000  2.380      0.743
#> 8   0.166  0.375  0.000  2.500      0.705
#> 9   0.007  0.081  0.000  1.000      0.610
#> 10  0.004  0.060  0.000  1.000      0.606
#> 11  0.204  0.425  0.000  2.400      0.603
#> 12  0.010  0.101  0.000  1.000      0.572
#> 13  0.004  0.063  0.000  1.000      0.534
#> 14  0.190  0.467  0.000  2.660      0.481
#> 15  0.007  0.083  0.000  1.000      0.441
#> 16  0.095  0.293  0.000  1.000      0.349
#> 17 34.002 13.444 13.913 88.736      0.345
#> 18  0.075  0.263  0.000  1.000      0.339
#> 19  0.794  0.474  0.200  3.000      0.286
#> 20  0.002  0.048  0.000  1.000      0.284
#> 21  0.147  0.354  0.000  1.000      0.252
#> 22  0.014  0.118  0.000  1.000      0.232
#> 23  0.199  0.399  0.000  1.000      0.231
#> 24  0.040  0.195  0.000  1.000      0.201
#> 25  0.001  0.032  0.000  1.000      0.192
#> 26  0.008  0.091  0.000  1.000      0.192
#> 27  0.023  0.151  0.000  1.000      0.190
#> 28  0.220  0.415  0.000  1.000      0.177
#> 29  0.032  0.177  0.000  1.000      0.176
#> 30  0.032  0.175  0.000  1.000      0.175
#> 31  0.037  0.190  0.000  1.000      0.155
#> 32  0.095  0.293  0.000  1.000      0.144
#> 33  0.026  0.159  0.000  1.000      0.144
#> 34  0.017  0.128  0.000  1.000      0.133
#> 35  0.055  0.227  0.000  1.000      0.121
#> 36  0.011  0.104  0.000  1.000      0.116
#> 37  0.007  0.081  0.000  1.000      0.108
#> 38  0.240  0.427  0.000  1.000      0.103
#> 39  0.151  0.358  0.000  1.000      0.101
#> 40  0.035  0.185  0.000  1.000      0.100
#> 41  0.012  0.110  0.000  1.000      0.096
#> 42  0.017  0.128  0.000  1.000      0.095
#> 43  0.185  0.388  0.000  1.000      0.087
#> 44  0.046  0.210  0.000  1.000      0.079
#> 45  0.049  0.216  0.000  1.000      0.073
#> 46  0.017  0.131  0.000  1.000      0.066
#> 47  0.019  0.135  0.000  1.000      0.063
#> 48  0.023  0.151  0.000  1.000      0.059
#> 49  0.015  0.122  0.000  1.000      0.047
#> 50  0.007  0.081  0.000  1.000      0.046
#> 51  0.019  0.135  0.000  1.000      0.038
#> 52  0.028  0.164  0.000  1.000      0.033
#> 53  0.188  0.391  0.000  1.000      0.032
#> 54  0.039  0.194  0.000  1.000      0.029
#> 55  0.011  0.103  0.000  1.000      0.026
#> 56  0.176  0.381  0.000  1.000      0.025
#> 57  0.043  0.202  0.000  1.000      0.024
#> 58  0.029  0.169  0.000  1.000      0.020
#> 59  0.025  0.157  0.000  1.000      0.020
#> 60  0.031  0.173  0.000  1.000      0.018
#> 61  0.011  0.104  0.000  1.000      0.016
#> 62  0.029  0.167  0.000  1.000      0.014
#> 63  0.016  0.127  0.000  1.000      0.012
#> 64  0.015  0.123  0.000  1.000      0.012
#> 65  0.042  0.200  0.000  1.000      0.010
#> 66  5.722  1.124  3.730  9.420      0.009